@@ -5,7 +5,6 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"reflect"
 	"time"
 
 	"google.golang.org/protobuf/types/known/timestamppb"
@@ -15,72 +14,45 @@ import (
 	"github.com/jmoiron/sqlx"
 	v1 "github.com/metal-stack/masterdata-api/api/v1"
 	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
 
 	// import for sqlx to use postgres driver
 	_ "github.com/lib/pq"
 )
 
-// exchangable for testing
+// exchangeable for testing
 var Now = time.Now
 
 // Storage is a interface to store objects.
-type Storage interface {
+type Storage[E Entity] interface {
 	// generic
-	Create(ctx context.Context, ve VersionedJSONEntity) error
-	Update(ctx context.Context, ve VersionedJSONEntity) error
-	Delete(ctx context.Context, ve VersionedJSONEntity) error
-	Get(ctx context.Context, id string, ve VersionedJSONEntity) error
-	GetHistory(ctx context.Context, id string, at time.Time, ve VersionedJSONEntity) error
-	Find(ctx context.Context, filter map[string]interface{}, paging *v1.Paging, result interface{}) (*uint64, error)
+	Create(ctx context.Context, ve E) error
+	Update(ctx context.Context, ve E) error
+	Delete(ctx context.Context, id string) error
+	Get(ctx context.Context, id string) (E, error)
+	GetHistory(ctx context.Context, id string, at time.Time, ve E) error
+	GetHistoryCreated(ctx context.Context, id string, ve E) error
+	Find(ctx context.Context, filter map[string]any, paging *v1.Paging) ([]E, *uint64, error)
 }
 
-const defaultPagingLimit = uint64(100)
-
-func addPaging(q squirrel.SelectBuilder, paging *v1.Paging) (squirrel.SelectBuilder, *uint64) {
-	if paging == nil {
-		return q, nil
-	}
-
-	limit := defaultPagingLimit
-	if paging.Count != nil {
-		limit = *paging.Count
-	}
-	offset := uint64(0)
-	nextpage := uint64(1)
-	if paging.Page != nil {
-		offset = *paging.Page * limit
-		nextpage = *paging.Page + 1
-	}
-	q = q.Limit(limit).Offset(offset)
-	return q, &nextpage
-}
-
-// JSONEntity is storable in json format
-type JSONEntity interface {
+// Entity defines a database entity which is stored in jsonb format and with version information
+type Entity interface {
 	JSONField() string
 	TableName() string
 	Schema() string
-}
-
-// VersionedEntity defines a database entity which is stored with version information
-type VersionedEntity interface {
 	GetMeta() *v1.Meta
 	Kind() string
 	APIVersion() string
 }
 
-// VersionedJSONEntity defines a database entity which is stored in jsonb format and with version information
-type VersionedJSONEntity interface {
-	JSONEntity
-	VersionedEntity
-}
-
-// Datastore is the adapter to talk to the database
-type Datastore struct {
-	log   *zap.Logger
-	db    *sqlx.DB
-	sb    squirrel.StatementBuilderType
-	types map[string]VersionedJSONEntity
+// datastore is the adapter to talk to the database
+type datastore[E Entity] struct {
+	log              *zap.Logger
+	db               *sqlx.DB
+	sb               squirrel.StatementBuilderType
+	jsonField        string
+	tableName        string
+	historyTableName string
 }
 
 type Op string
@@ -92,75 +64,78 @@ const (
 )
 
 // NewPostgresStorage creates a new Storage which uses postgres.
-func NewPostgresStorage(logger *zap.Logger, host, port, user, password, dbname, sslmode string, ves ...VersionedJSONEntity) (*Datastore, error) {
+func NewPostgresDB(logger *zap.Logger, host, port, user, password, dbname, sslmode string, ves ...Entity) (*sqlx.DB, error) {
 	db, err := sqlx.Connect("postgres", fmt.Sprintf("host=%s port=%s user=%s dbname=%s password=%s sslmode=%s", host, port, user, dbname, password, sslmode))
 	if err != nil {
 		return nil, fmt.Errorf("unable to connect to database: %w", err)
 	}
-	types := make(map[string]VersionedJSONEntity)
 	for _, ve := range ves {
 		jsonField := ve.JSONField()
 		logger.Info("creating schema", zap.String("entity", jsonField))
-		_, err = db.Exec(ve.Schema())
+		_, err := db.Exec(ve.Schema())
 		if err != nil {
 			logger.Fatal("unable to create schema", zap.String("entity", jsonField), zap.Error(err))
 			return nil, err
 		}
-		types[jsonField] = ve
 	}
-	ds := &Datastore{
-		log:   logger,
-		db:    db,
-		sb:    squirrel.StatementBuilder.PlaceholderFormat(squirrel.Dollar).RunWith(db),
-		types: types,
+	return db, nil
+}
+
+// New creates a new Storage which uses the given database abstraction.
+func New[E Entity](logger *zap.Logger, db *sqlx.DB, e E) (Storage[E], error) {
+	ds := &datastore[E]{
+		log:              logger,
+		db:               db,
+		sb:               squirrel.StatementBuilder.PlaceholderFormat(squirrel.Dollar).RunWith(db),
+		jsonField:        e.JSONField(),
+		tableName:        e.TableName(),
+		historyTableName: fmt.Sprintf("%s_history", e.TableName()),
 	}
 	return ds, nil
 }
 
 // Create a entity
-func (ds *Datastore) Create(ctx context.Context, ve VersionedJSONEntity) error {
-	jsonField := ve.JSONField()
-	tableName := ve.TableName()
-	_, ok := ds.types[jsonField]
-	if !ok {
-		return fmt.Errorf("type:%s is not registered", jsonField)
-	}
+func (ds *datastore[E]) Create(ctx context.Context, ve E) error {
+	ds.log.Debug("create", zap.String("entity", ds.jsonField), zap.Any("value", ve))
 	meta := ve.GetMeta()
 	if meta == nil {
-		return fmt.Errorf("create of type:%s failed, meta is nil", jsonField)
+		return fmt.Errorf("create of type:%s failed, meta is nil", ds.jsonField)
 	}
 	id := meta.GetId()
 	if id == "" {
-		id = uuid.Must(uuid.NewRandom()).String()
+		id = uuid.NewString()
 		meta.SetId(id)
 	}
 	kind := meta.GetKind()
 	if kind == "" {
 		meta.Kind = ve.Kind()
 	} else if kind != ve.Kind() {
-		return fmt.Errorf("create of type:%s failed, kind is set to:%s but must be:%s", jsonField, kind, ve.Kind())
+		return fmt.Errorf("create of type:%s failed, kind is set to:%s but must be:%s", ds.jsonField, kind, ve.Kind())
 	}
 	apiVersion := meta.GetApiversion()
 	if apiVersion == "" {
 		meta.Apiversion = ve.APIVersion()
 	} else if apiVersion != ve.APIVersion() {
-		return fmt.Errorf("create of type:%s failed, apiversion must be set to:%s", jsonField, ve.APIVersion())
+		return fmt.Errorf("create of type:%s failed, apiversion must be set to:%s", ds.jsonField, ve.APIVersion())
 	}
 
-	createdAtPb, createdAt := PbNow()
+	createdAtPb, createdAt := pbNow()
 	meta.SetVersion(0)
 	meta.SetCreatedTime(createdAtPb)
 
 	q := ds.sb.Insert(
-		tableName,
-	).SetMap(map[string]interface{}{
-		"id":      id,
-		jsonField: ve,
+		ds.tableName,
+	).SetMap(map[string]any{
+		"id":         id,
+		ds.jsonField: ve,
 	}).Suffix(
-		"RETURNING " + jsonField,
+		"RETURNING " + ds.jsonField,
 	)
-	sql, vals, _ := q.ToSql()
-	ds.log.Info("create", zap.String("entity", tableName), zap.String("sql", sql), zap.Any("values", vals))
+
+	if ce := ds.log.Check(zapcore.DebugLevel, ""); ce != nil {
+		sql, vals, _ := q.ToSql()
+		ds.log.Debug("create", zap.String("entity", ds.jsonField), zap.String("sql", sql), zap.Any("values", vals))
+	}
 
 	tx, err := ds.db.BeginTx(ctx, nil)
 	if err != nil {
@@ -171,7 +146,7 @@ func (ds *Datastore) Create(ctx context.Context, ve VersionedJSONEntity) error {
 	err = q.RunWith(tx).QueryRowContext(ctx).Scan(ve)
 	if err != nil {
 		if IsErrorCode(err, UniqueViolationError) {
-			return NewDuplicateKeyError(fmt.Sprintf("an entity of type:%s with the id:%s already exists", jsonField, meta.Id))
+			return NewDuplicateKeyError(fmt.Sprintf("an entity of type:%s with the id:%s already exists", ds.jsonField, meta.Id))
 		}
 		return err
 	}
@@ -183,74 +158,66 @@ func (ds *Datastore) Create(ctx context.Context, ve VersionedJSONEntity) error {
 }
 
 // Update the entity
-func (ds *Datastore) Update(ctx context.Context, ve VersionedJSONEntity) error {
-	jsonField := ve.JSONField()
-	tableName := ve.TableName()
-	_, ok := ds.types[jsonField]
-	if !ok {
-		return fmt.Errorf("type:%s is not registered", jsonField)
-	}
+func (ds *datastore[E]) Update(ctx context.Context, ve E) error {
+	ds.log.Info("update", zap.String("entity", ds.jsonField))
 	meta := ve.GetMeta()
 	if meta == nil {
-		return fmt.Errorf("update of type:%s failed, meta is nil", jsonField)
+		return fmt.Errorf("update of type:%s failed, meta is nil", ds.jsonField)
 	}
 	id := meta.GetId()
 	if id == "" {
-		return fmt.Errorf("entity of type:%s has no id, cannot update: %v", jsonField, ve)
+		return fmt.Errorf("entity of type:%s has no id, cannot update: %v", ds.jsonField, ve)
 	}
 	kind := meta.GetKind()
 	if kind == "" {
 		meta.Kind = ve.Kind()
 	} else if kind != ve.Kind() {
-		return fmt.Errorf("update of type:%s failed, kind is set to:%s but must be:%s", jsonField, kind, ve.Kind())
+		return fmt.Errorf("update of type:%s failed, kind is set to:%s but must be:%s", ds.jsonField, kind, ve.Kind())
 	}
 	apiVersion := meta.GetApiversion()
 	if apiVersion == "" {
 		meta.Apiversion = ve.APIVersion()
 	} else if apiVersion != ve.APIVersion() {
-		return fmt.Errorf("update of type:%s failed, apiversion must be set to:%s", jsonField, ve.APIVersion())
+		return fmt.Errorf("update of type:%s failed, apiversion must be set to:%s", ds.jsonField, ve.APIVersion())
 	}
 
-	elemt := reflect.TypeOf(ve).Elem()
-	existingVE, ok := reflect.New(elemt).Interface().(VersionedJSONEntity)
-	if !ok {
-		return fmt.Errorf("entity is not a VersionedJSONEntity: %v", ve)
-	}
-
-	err := ds.Get(ctx, id, existingVE)
+	existingVE, err := ds.Get(ctx, id)
 	if err != nil {
-		return fmt.Errorf("update - no entity of type:%s with id:%s found", jsonField, id)
+		return fmt.Errorf("update - no entity of type:%s with id:%s found", ds.jsonField, id)
 	}
 
 	if ve.GetMeta().GetVersion() < existingVE.GetMeta().GetVersion() {
 		return NewOptimisticLockError(
 			fmt.Sprintf("optimistic lock error updating %s with id %s, existing version %d mismatches entity version %d",
-				jsonField, id, existingVE.GetMeta().GetVersion(), ve.GetMeta().GetVersion(),
+				ds.jsonField, id, existingVE.GetMeta().GetVersion(), ve.GetMeta().GetVersion(),
 			),
 		)
 	}
 
-	pbNow, now := PbNow()
+	pbNow, now := pbNow()
 
 	ve.GetMeta().SetVersion(ve.GetMeta().GetVersion() + 1)
 	ve.GetMeta().SetUpdatedTime(pbNow)
 
-	// handle non updateable fields like created_time
+	// handle non updatable fields like created_time
 	// simple strategy: copy unmodifiable fields from existing before update
 	ve.GetMeta().SetCreatedTime(existingVE.GetMeta().GetCreatedTime())
 
-	q := ds.sb.Update(tableName).
-		SetMap(map[string]interface{}{
-			jsonField: ve,
+	q := ds.sb.Update(ds.tableName).
+		SetMap(map[string]any{
+			ds.jsonField: ve,
 		}).
 		Where(squirrel.Eq{
 			"id": id,
 		}).
 		Suffix(
-			"RETURNING " + jsonField,
+			"RETURNING " + ds.jsonField,
 		)
-	sql, vals, _ := q.ToSql()
-	ds.log.Info("update", zap.String("entity", tableName), zap.String("sql", sql), zap.Any("values", vals))
+
+	if ce := ds.log.Check(zapcore.DebugLevel, ""); ce != nil {
+		sql, vals, _ := q.ToSql()
+		ds.log.Debug("update", zap.String("entity", ds.jsonField), zap.String("sql", sql), zap.Any("values", vals))
+	}
 
 	tx, err := ds.db.BeginTx(ctx, nil)
 	if err != nil {
@@ -274,66 +241,40 @@ func (ds *Datastore) Update(ctx context.Context, ve VersionedJSONEntity) error {
 
 // Get the entity for given id
 // returns NotFoundError if no entity can be found
-func (ds *Datastore) Get(ctx context.Context, id string, ve VersionedJSONEntity) error {
-	jsonField := ve.JSONField()
-	tableName := ve.TableName()
-	_, ok := ds.types[jsonField]
-	if !ok {
-		return fmt.Errorf("type:%s is not registered", jsonField)
-	}
+func (ds *datastore[E]) Get(ctx context.Context, id string) (E, error) {
+	ds.log.Debug("get", zap.String("entity", ds.jsonField), zap.String("id", id))
+	var zero E
 	q := ds.sb.Select(
-		jsonField,
+		ds.jsonField,
 	).From(
-		tableName,
+		ds.tableName,
 	).Where(squirrel.Eq{
 		"id": id,
 	})
 
-	sql, _, _ := q.ToSql()
-	ds.log.Info("get", zap.String("entity", jsonField), zap.String("sql", sql), zap.String("id", id))
-	rows, err := q.QueryContext(ctx)
+	row := q.QueryRowContext(ctx)
+	e := new(E)
+	err := row.Scan(e)
 	if err != nil {
-		return err
+		if errors.Is(err, sql.ErrNoRows) {
+			return zero, NewNotFoundError(fmt.Sprintf("%s with id:%s not found %v", ds.jsonField, id, err))
+		}
+		return zero, err
 	}
-	defer func() {
-		_ = rows.Close()
-		_ = rows.Err()
-	}()
-	if rows.Next() {
-		return rows.Scan(ve)
-	}
-	// we have no row
-	return NewNotFoundError(fmt.Sprintf("entity of type:%s with id:%s not found", jsonField, id))
+	return *e, nil
 }
 
 // Delete deletes the entity
-func (ds *Datastore) Delete(ctx context.Context, ve VersionedJSONEntity) error {
-	jsonField := ve.JSONField()
-	tableName := ve.TableName()
-	_, ok := ds.types[jsonField]
-	if !ok {
-		return fmt.Errorf("type:%s is not registered", jsonField)
-	}
-
-	elemt := reflect.TypeOf(ve).Elem()
-	existingVE, ok := reflect.New(elemt).Interface().(VersionedJSONEntity)
-	if !ok {
-		return fmt.Errorf("entity is not a VersionedJSONEntity: %v", ve)
-	}
-	err := ds.Get(ctx, ve.GetMeta().Id, existingVE)
+func (ds *datastore[E]) Delete(ctx context.Context, id string) error {
+	ds.log.Debug("delete", zap.String("entity", ds.jsonField), zap.String("id", id))
+	ve, err := ds.Get(ctx, id)
 	if err != nil {
 		return err
 	}
 
 	// delete dataset in table
-	q := ds.sb.Delete(tableName).
-		Where(squirrel.Eq{"id": ve.GetMeta().GetId()})
-	sql, _, err := q.ToSql()
-	if err != nil {
-		return err
-	}
-	ds.log.Debug("delete", zap.String("entity", jsonField), zap.String("sql", sql))
-
+	q := ds.sb.Delete(ds.tableName).
+		Where(squirrel.Eq{"id": id})
 	// in tx
 	tx, err := ds.db.BeginTx(ctx, nil)
 	if err != nil {
@@ -351,14 +292,14 @@ func (ds *Datastore) Delete(ctx context.Context, ve VersionedJSONEntity) error {
 		return err
 	}
 	if rowsAffected > 1 {
-		return NewDataCorruptionError(fmt.Sprintf("datacorruption: delete of %s with id %s affected %d rows", jsonField, ve.GetMeta().Id, rowsAffected))
+		return NewDataCorruptionError(fmt.Sprintf("data corruption: delete of %s with id %s affected %d rows", ds.jsonField, id, rowsAffected))
 	}
 	if rowsAffected < 1 {
-		return NewNotFoundError(fmt.Sprintf("not found: delete of %s with id %s affected %d rows", jsonField, ve.GetMeta().Id, rowsAffected))
+		return NewNotFoundError(fmt.Sprintf("not found: delete of %s with id %s affected %d rows", ds.jsonField, id, rowsAffected))
 	}
 
 	// insert dataset in history table
-	err = ds.insertHistory(existingVE, opDelete, Now(), tx)
+	err = ds.insertHistory(ve, opDelete, Now(), tx)
 	if err != nil {
 		return err
 	}
@@ -367,28 +308,10 @@ func (ds *Datastore) Delete(ctx context.Context, ve VersionedJSONEntity) error {
 }
 
 // Find returns matching elements from the database
-func (ds *Datastore) Find(ctx context.Context, filter map[string]interface{}, paging *v1.Paging, result interface{}) (*uint64, error) {
-	resultv := reflect.ValueOf(result)
-	if resultv.Kind() != reflect.Ptr || resultv.Elem().Kind() != reflect.Slice {
-		return nil, fmt.Errorf("result argument must be a slice address")
-	}
-
-	slicev := resultv.Elem()
-	elemt := slicev.Type().Elem()
-
-	ve, ok := reflect.New(elemt).Interface().(VersionedJSONEntity)
-	if !ok {
-		return nil, fmt.Errorf("result slice element type must implement VersionedJSONEntity-Interface")
-	}
-	jsonField := ve.JSONField()
-	tableName := ve.TableName()
-	_, ok = ds.types[jsonField]
-	if !ok {
-		return nil, fmt.Errorf("type:%s is not registered", jsonField)
-	}
-
-	q := ds.sb.Select(jsonField).
-		From(tableName)
+func (ds *datastore[E]) Find(ctx context.Context, filter map[string]any, paging *v1.Paging) ([]E, *uint64, error) {
+	ds.log.Debug("find", zap.String("entity", ds.jsonField), zap.Any("filter", filter))
+	q := ds.sb.Select(ds.jsonField).
+		From(ds.tableName)
 
 	if len(filter) > 0 {
 		q = q.Where(filter)
@@ -398,43 +321,42 @@ func (ds *Datastore) Find(ctx context.Context, filter map[string]interface{}, pa
 	// Add paging query if paging is defined
 	q, nextPage := addPaging(q, paging)
 
-	sql, vals, _ := q.ToSql()
-	ds.log.Debug("find", zap.String("sql", sql), zap.Any("values", vals))
+	if ce := ds.log.Check(zapcore.DebugLevel, ""); ce != nil {
+		sql, vals, _ := q.ToSql()
+		ds.log.Debug("find", zap.String("entity", ds.jsonField), zap.String("sql", sql), zap.Any("values", vals))
+	}
 
 	rows, err := q.QueryContext(ctx)
 	if err != nil {
-		return nil, err
+		return nil, nil, err
 	}
 	defer func() {
 		_ = rows.Close()
-		_ = rows.Err()
 	}()
 
-	rowcount := uint64(0)
+	ves := []E{}
 	for rows.Next() {
-		elemp := reflect.New(elemt)
-		err = rows.Scan(elemp.Interface())
+		e := new(E)
+		err = rows.Scan(e)
 		if err != nil {
-			return nil, err
+			return nil, nil, err
 		}
-		slicev = reflect.Append(slicev, elemp.Elem())
-		rowcount++
+		ves = append(ves, *e)
 	}
-	resultv.Elem().Set(slicev)
 
 	err = rows.Err()
 	if err != nil {
-		return nil, err
+		return nil, nil, err
 	}
-	if paging != nil && paging.Count != nil && rowcount == *paging.Count {
-		return nextPage, err
+	if paging != nil && paging.Count != nil && uint64(len(ves)) == *paging.Count {
+		return ves, nextPage, err
 	}
-	return nil, err
+	return ves, nil, nil
 }
 
 // Get the history entity for given id and latest before or equal the given point in time
 // returns NotFoundError if no entity can be found
-func (ds *Datastore) GetHistory(ctx context.Context, id string, at time.Time, ve VersionedJSONEntity) error {
+func (ds *datastore[E]) GetHistory(ctx context.Context, id string, at time.Time, ve E) error {
 	return ds.getHistoryWithPredicate(ctx, squirrel.And{
 		squirrel.Eq{
 			"id": id,
@@ -446,7 +368,7 @@ func (ds *Datastore) GetHistory(ctx context.Context, id string, at time.Time, ve
 }
 
 // Get the first history entity for given id, returns NotFoundError if no entity can be found
-func (ds *Datastore) GetHistoryCreated(ctx context.Context, id string, ve VersionedJSONEntity) error {
+func (ds *datastore[E]) GetHistoryCreated(ctx context.Context, id string, ve E) error {
 	return ds.getHistoryWithPredicate(ctx, squirrel.And{
 		squirrel.Eq{
 			"id": id,
@@ -459,17 +381,11 @@ func (ds *Datastore) GetHistoryCreated(ctx context.Context, id string, ve Versio
 
 // Get the top matching history entity for given filter criteria,
 // returns NotFoundError if no entity can be found
-func (ds *Datastore) getHistoryWithPredicate(ctx context.Context, pred interface{}, ve VersionedJSONEntity) error {
-	jsonField := ve.JSONField()
-	tableName := historyTablename(ve.TableName())
-	_, ok := ds.types[jsonField]
-	if !ok {
-		return fmt.Errorf("type:%s is not registered", jsonField)
-	}
-	q := ds.sb.Select(jsonField).From(tableName).Where(pred).OrderByClause("created_at DESC").Limit(1)
+func (ds *datastore[E]) getHistoryWithPredicate(ctx context.Context, pred any, ve E) error {
+	q := ds.sb.Select(ds.jsonField).From(ds.historyTableName).Where(pred).OrderByClause("created_at DESC").Limit(1)
 
 	sql, _, _ := q.ToSql()
-	ds.log.Info("get", zap.String("entity", jsonField), zap.String("sql", sql), zap.Any("predicate", pred))
+	ds.log.Info("get", zap.String("entity", ds.jsonField), zap.String("sql", sql), zap.Any("predicate", pred))
 	rows, err := q.QueryContext(ctx)
 	if err != nil {
 		return err
@@ -483,19 +399,17 @@ func (ds *Datastore) getHistoryWithPredicate(ctx context.Context, pred interface
 		return rows.Scan(ve)
 	}
 	// we have no row
-	return NewNotFoundError(fmt.Sprintf("entity of type:%s with predicate:%s not found", jsonField, pred))
+	return NewNotFoundError(fmt.Sprintf("entity of type:%s with predicate:%s not found", ds.jsonField, pred))
 }
 
 // insertHistory inserts the given entity in the history table of the entity using the runner, which may be a Tx.
-func (ds *Datastore) insertHistory(ve VersionedJSONEntity, op Op, createdAt time.Time, runner squirrel.BaseRunner) error {
-	jsonField := ve.JSONField()
-	tableName := ve.TableName()
-	qh := ds.sb.Insert(historyTablename(tableName)).
-		SetMap(map[string]interface{}{
+func (ds *datastore[E]) insertHistory(ve E, op Op, createdAt time.Time, runner squirrel.BaseRunner) error {
+	qh := ds.sb.Insert(ds.historyTableName).
+		SetMap(map[string]any{
 			"id":         ve.GetMeta().Id,
 			"op":         op,
 			"created_at": createdAt,
-			jsonField:    ve,
+			ds.jsonField: ve,
 		})
 	_, err := qh.RunWith(runner).Exec()
 	if err != nil {
@@ -504,22 +418,38 @@ func (ds *Datastore) insertHistory(ve VersionedJSONEntity, op Op, createdAt time
 	return nil
 }
 
-// historyTablename returns the tablename of the table-twin with historic data.
-func historyTablename(table string) string {
-	return fmt.Sprintf("%s_history", table)
-}
-
-// PbNow returns the current time as Protobuf and time
-func PbNow() (*timestamppb.Timestamp, time.Time) {
+// pbNow returns the current time as Protobuf and time
+func pbNow() (*timestamppb.Timestamp, time.Time) {
 	now := Now()
 	nowPb := timestamppb.New(now)
 	return nowPb, now
 }
 
 // rollback tries to rollback the given transaction and logs an eventual rollback error
-func (ds *Datastore) rollback(tx *sql.Tx) {
+func (ds *datastore[E]) rollback(tx *sql.Tx) {
 	err := tx.Rollback()
 	if err != nil && !errors.Is(err, sql.ErrTxDone) {
 		ds.log.Error("error rolling back", zap.Error(err))
 	}
+}
+
+const defaultPagingLimit = uint64(100)
+
+func addPaging(q squirrel.SelectBuilder, paging *v1.Paging) (squirrel.SelectBuilder, *uint64) {
+	if paging == nil {
+		return q, nil
+	}
+
+	limit := defaultPagingLimit
+	if paging.Count != nil {
+		limit = *paging.Count
+	}
+	offset := uint64(0)
+	nextpage := uint64(1)
+	if paging.Page != nil {
+		offset = *paging.Page * limit
+		nextpage = *paging.Page + 1
+	}
+	q = q.Limit(limit).Offset(offset)
+	return q, &nextpage
 }
