@@ -8,30 +8,20 @@ import (
 	"testing"
 	"time"
 
+	"github.com/jmoiron/sqlx"
 	"github.com/testcontainers/testcontainers-go"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
 	"github.com/stretchr/testify/require"
 
 	v1 "github.com/metal-stack/masterdata-api/api/v1"
+	"github.com/metal-stack/metal-lib/pkg/pointer"
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
+	"go.uber.org/zap/zaptest"
 )
 
-var (
-	ds  *Datastore
-	log *zap.Logger
-)
-
-// to test unregistered type checks
-type invalidVersionedEntity struct{}
-
-func (v *invalidVersionedEntity) JSONField() string  { return "invalid" }
-func (v *invalidVersionedEntity) TableName() string  { return "" }
-func (v *invalidVersionedEntity) Schema() string     { return "" }
-func (v *invalidVersionedEntity) Kind() string       { return "Invalid" }
-func (v *invalidVersionedEntity) APIVersion() string { return "v1" }
-func (v *invalidVersionedEntity) GetMeta() *v1.Meta  { return nil }
+var db *sqlx.DB
 
 func TestMain(m *testing.M) {
 	code := 0
@@ -47,63 +37,18 @@ func TestMain(m *testing.M) {
 		}
 	}()
 
-	log, _ = zap.NewProduction()
-
-	ctx := context.Background()
-	req := testcontainers.ContainerRequest{
-		Image:        "postgres:12-alpine",
-		ExposedPorts: []string{"5432/tcp"},
-		Env:          map[string]string{"POSTGRES_PASSWORD": "password"},
-		// TODO: should work, but dont, hence using the loop below to check pg is up an ready for connections.
-		// WaitingFor:   wait.ForLog("database system is ready to accept connections"),
-		// WaitingFor: wait.ForListeningPort("5432/tcp"),
-	}
-
-	postgres, err := testcontainers.GenericContainer(ctx, testcontainers.GenericContainerRequest{
-		ContainerRequest: req,
-		Started:          true,
-	})
+	var err error
+	db, err = createPostgresConnection()
 	if err != nil {
-		log.Info(err.Error())
+		panic(err)
 	}
-	defer func() {
-		err := postgres.Terminate(ctx)
-		if err != nil {
-			log.Fatal(err.Error())
-		}
-	}()
-
-	ip, err := postgres.Host(ctx)
-	if err != nil {
-		log.Fatal(err.Error())
-	}
-	port, err := postgres.MappedPort(ctx, "5432/tcp")
-	if err != nil {
-		log.Fatal(err.Error())
-	}
-
-	for {
-		var err error
-		ds, err = NewPostgresStorage(log, ip, port.Port(), "postgres", "password", "postgres", "disable", &v1.Project{}, &v1.Tenant{})
-		if err != nil {
-			time.Sleep(100 * time.Millisecond)
-			continue
-		}
-		err = ds.db.Ping()
-		if err != nil {
-			log.Error("Could not connect to postgres server", zap.Error(err))
-		}
-		if err == nil {
-			break
-		}
-	}
-
-	log.Info("connected to postgres")
 	code = m.Run()
 }
 
 func TestCRUD(t *testing.T) {
-	assert.NotNil(t, ds, "Datastore must not be nil")
+	tenantDS, err := New(zaptest.NewLogger(t), db, &v1.Tenant{})
+	require.NoError(t, err)
+	assert.NotNil(t, tenantDS, "Datastore must not be nil")
 	ctx := context.Background()
 	tcr := &v1.Tenant{
 		Meta:        &v1.Meta{Id: "tenant-1"},
@@ -111,7 +56,7 @@ func TestCRUD(t *testing.T) {
 		Description: "A very important Tenant",
 	}
 
-	err := ds.Create(ctx, tcr)
+	err = tenantDS.Create(ctx, tcr)
 	assert.NoError(t, err)
 	assert.NotNil(t, tcr)
 	// specified id is persisted
@@ -121,12 +66,11 @@ func TestCRUD(t *testing.T) {
 	assert.Equal(t, "A Tenant", tcr.GetName())
 	assert.Equal(t, "A very important Tenant", tcr.GetDescription())
 
-	err = ds.Create(ctx, tcr)
+	err = tenantDS.Create(ctx, tcr)
 	assert.EqualError(t, err, "an entity of type:tenant with the id:tenant-1 already exists")
 
 	// get existing
-	var tgr v1.Tenant
-	err = ds.Get(ctx, tcr.Meta.GetId(), &tgr)
+	tgr, err := tenantDS.Get(ctx, tcr.Meta.GetId())
 	assert.NoError(t, err)
 	assert.NotNil(t, &tgr)
 	assert.Equal(t, "tenant-1", tgr.Meta.Id)
@@ -134,14 +78,13 @@ func TestCRUD(t *testing.T) {
 	assert.Equal(t, "A very important Tenant", tgr.GetDescription())
 
 	// get unknown
-	var tgr2 v1.Tenant
-	err = ds.Get(ctx, "unknown-id", &tgr2)
+	tgr2, err := tenantDS.Get(ctx, "unknown-id")
 	assert.Error(t, err)
-	assert.EqualError(t, err, "entity of type:tenant with id:unknown-id not found")
+	assert.EqualError(t, err, "tenant with id:unknown-id not found sql: no rows in result set")
 	assert.NotNil(t, &tgr2)
 
 	// update without meta and id
-	err = ds.Update(ctx, &tgr2)
+	err = tenantDS.Update(ctx, tgr2)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "update of type:tenant failed, meta is nil")
 
@@ -151,41 +94,42 @@ func TestCRUD(t *testing.T) {
 		Name:        "A second Tenant",
 		Description: "A not so important Tenant",
 	}
-	err = ds.Update(ctx, tcr2)
+	err = tenantDS.Update(ctx, tcr2)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "update - no entity of type:tenant with id:tenant-2 found")
 
 	// update name
 	tcr.Name = "Important Tenant"
-	err = ds.Update(ctx, tcr)
+	err = tenantDS.Update(ctx, tcr)
 	assert.NoError(t, err)
 	assert.Equal(t, "Important Tenant", tcr.GetName())
 
 	// find existing
-	var tenants []v1.Tenant
-	filter := make(map[string]interface{})
+	filter := make(map[string]any)
 	// filter["tenant->>name"] = "Important Tenant"
 	filter["id"] = "tenant-1"
-	_, err = ds.Find(ctx, filter, nil, &tenants)
+	tenants, _, err := tenantDS.Find(ctx, filter, nil)
 	assert.NoError(t, err)
 	assert.NotNil(t, tenants)
 	assert.Len(t, tenants, 1)
 	assert.Equal(t, "Important Tenant", tenants[0].Name)
 
 	// delete existing
-	err = ds.Delete(ctx, tcr)
+	err = tenantDS.Delete(ctx, tcr.Meta.Id)
 	assert.NoError(t, err)
 	assert.NotNil(t, tcr)
 
 	// delete not existing
-	err = ds.Delete(ctx, tcr)
+	err = tenantDS.Delete(ctx, tcr.Meta.Id)
 	assert.Error(t, err)
-	assert.EqualError(t, err, "entity of type:tenant with id:tenant-1 not found")
+	assert.EqualError(t, err, "tenant with id:tenant-1 not found sql: no rows in result set")
 
 }
 
 func TestUpdateOptimisticLock(t *testing.T) {
-	assert.NotNil(t, ds, "Datastore must not be nil")
+	tenantDS, err := New(zaptest.NewLogger(t), db, &v1.Tenant{})
+	require.NoError(t, err)
+	assert.NotNil(t, tenantDS, "Datastore must not be nil")
 	ctx := context.Background()
 	tcr := &v1.Tenant{
 		Meta:        &v1.Meta{Id: "tenant-2"},
@@ -193,30 +137,28 @@ func TestUpdateOptimisticLock(t *testing.T) {
 		Description: "A very important Tenant",
 	}
 
-	err := ds.Create(ctx, tcr)
+	err = tenantDS.Create(ctx, tcr)
 	assert.NoError(t, err)
 	assert.NotNil(t, tcr)
 	assert.Equal(t, int64(0), tcr.Meta.Version)
 	assert.Equal(t, "A Tenant", tcr.GetName())
 
 	// get from db
-	tget := &v1.Tenant{}
-	err = ds.Get(ctx, tcr.Meta.Id, tget)
+	tget, err := tenantDS.Get(ctx, tcr.Meta.Id)
 	require.NoError(t, err)
 	require.Equal(t, tcr.Meta.Id, tget.Meta.Id)
 	assert.Equal(t, int64(0), tget.Meta.Version)
 
 	// update instance
 	tcr.Name = "updated name"
-	err = ds.Update(ctx, tcr)
+	err = tenantDS.Update(ctx, tcr)
 	assert.NoError(t, err)
 	assert.NotNil(t, tcr)
 	// incremented version after update
 	assert.Equal(t, int64(1), tcr.Meta.Version)
 
 	// re-read from db
-	var tgr v1.Tenant
-	err = ds.Get(ctx, tcr.Meta.GetId(), &tgr)
+	tgr, err := tenantDS.Get(ctx, tcr.Meta.GetId())
 	assert.NoError(t, err)
 	// version is incremented
 	assert.Equal(t, int64(1), tgr.Meta.Version)
@@ -225,20 +167,16 @@ func TestUpdateOptimisticLock(t *testing.T) {
 
 	// try to update older version --> optimistic lock error
 	tget.Name = "updated older entity"
-	err = ds.Update(ctx, tget)
+	err = tenantDS.Update(ctx, tget)
 	require.Equal(t, err, NewOptimisticLockError(fmt.Sprintf("optimistic lock error updating tenant with id %s, existing version 1 mismatches entity version 0", tget.GetMeta().Id)))
 }
 
 func TestCreate(t *testing.T) {
 	const t1 = "t1"
-	assert.NotNil(t, ds, "Datastore must not be nil")
+	tenantDS, err := New(zaptest.NewLogger(t), db, &v1.Tenant{})
+	require.NoError(t, err)
+	assert.NotNil(t, tenantDS, "Datastore must not be nil")
 	ctx := context.Background()
-	// unregistered type
-
-	ive := &invalidVersionedEntity{}
-	err := ds.Create(ctx, ive)
-	assert.Error(t, err)
-	assert.EqualError(t, err, "type:invalid is not registered")
 
 	tcr1 := &v1.Tenant{
 		Name:        "atenant",
@@ -246,7 +184,7 @@ func TestCreate(t *testing.T) {
 	}
 
 	// meta is nil
-	err = ds.Create(ctx, tcr1)
+	err = tenantDS.Create(ctx, tcr1)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "create of type:tenant failed, meta is nil")
 
@@ -256,7 +194,7 @@ func TestCreate(t *testing.T) {
 		Name:        "atenant",
 		Description: "A Tenant",
 	}
-	err = ds.Create(ctx, tcr1)
+	err = tenantDS.Create(ctx, tcr1)
 	assert.NoError(t, err)
 	// specified id is persisted
 	assert.Equal(t, t1, tcr1.Meta.Id)
@@ -271,7 +209,7 @@ func TestCreate(t *testing.T) {
 		Name:        "btenant",
 		Description: "B Tenant",
 	}
-	err = ds.Create(ctx, tcr2)
+	err = tenantDS.Create(ctx, tcr2)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "an entity of type:tenant with the id:t1 already exists")
 
@@ -281,7 +219,7 @@ func TestCreate(t *testing.T) {
 		Name:        "ctenant",
 		Description: "C Tenant",
 	}
-	err = ds.Create(ctx, tcr3)
+	err = tenantDS.Create(ctx, tcr3)
 	assert.NoError(t, err)
 	assert.NotNil(t, tcr3.GetMeta().GetId())
 	assert.NotEmpty(t, tcr3.GetMeta().GetId())
@@ -294,7 +232,7 @@ func TestCreate(t *testing.T) {
 		Name:        "dtenant",
 		Description: "D Tenant",
 	}
-	err = ds.Create(ctx, tcr4)
+	err = tenantDS.Create(ctx, tcr4)
 	assert.NoError(t, err)
 	assert.NotNil(t, tcr3.GetMeta().GetApiversion())
 	assert.NotEmpty(t, tcr3.GetMeta().GetApiversion())
@@ -309,7 +247,7 @@ func TestCreate(t *testing.T) {
 		Name:        "etenant",
 		Description: "E Tenant",
 	}
-	err = ds.Create(ctx, tcr5)
+	err = tenantDS.Create(ctx, tcr5)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "create of type:tenant failed, kind is set to:Project but must be:Tenant")
 
@@ -319,28 +257,24 @@ func TestCreate(t *testing.T) {
 		Name:        "ftenant",
 		Description: "F Tenant",
 	}
-	err = ds.Create(ctx, tcr6)
+	err = tenantDS.Create(ctx, tcr6)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "create of type:tenant failed, apiversion must be set to:v1")
 }
 
 func TestUpdate(t *testing.T) {
 	const t3 = "t3"
-	assert.NotNil(t, ds, "Datastore must not be nil")
+	tenantDS, err := New(zaptest.NewLogger(t), db, &v1.Tenant{})
+	require.NoError(t, err)
+	assert.NotNil(t, tenantDS, "Datastore must not be nil")
 	ctx := context.Background()
-	// unregistered type
-
-	ive := &invalidVersionedEntity{}
-	err := ds.Update(ctx, ive)
-	assert.Error(t, err)
-	assert.EqualError(t, err, "type:invalid is not registered")
 
 	// meta is nil
 	tcr1 := &v1.Tenant{
 		Name:        "ctenant",
 		Description: "C Tenant",
 	}
-	err = ds.Update(ctx, tcr1)
+	err = tenantDS.Update(ctx, tcr1)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "update of type:tenant failed, meta is nil")
 
@@ -350,9 +284,9 @@ func TestUpdate(t *testing.T) {
 		Name:        "ctenant",
 		Description: "C Tenant",
 	}
-	err = ds.Update(ctx, tcr1)
+	err = tenantDS.Update(ctx, tcr1)
 	assert.Error(t, err)
-	assert.EqualError(t, err, "entity of type:tenant has no id, cannot update: meta:{} name:\"ctenant\" description:\"C Tenant\"")
+	assert.ErrorContains(t, err, "entity of type:tenant has no id, cannot update: meta:{}")
 
 	// tenant with id is not found
 	tcr1 = &v1.Tenant{
@@ -360,11 +294,11 @@ func TestUpdate(t *testing.T) {
 		Name:        "ctenant",
 		Description: "C Tenant",
 	}
-	err = ds.Update(ctx, tcr1)
+	err = tenantDS.Update(ctx, tcr1)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "update - no entity of type:tenant with id:t3 found")
 	// create tenant
-	err = ds.Create(ctx, tcr1)
+	err = tenantDS.Create(ctx, tcr1)
 	assert.NoError(t, err)
 	assert.Equal(t, t3, tcr1.GetMeta().GetId())
 	assert.Equal(t, "ctenant", tcr1.GetName())
@@ -375,7 +309,7 @@ func TestUpdate(t *testing.T) {
 
 	// now update existing
 	tcr1.Description = "C Tenant 3"
-	err = ds.Update(ctx, tcr1)
+	err = tenantDS.Update(ctx, tcr1)
 	assert.NoError(t, err)
 	assert.Equal(t, t3, tcr1.GetMeta().GetId())
 	assert.Equal(t, "ctenant", tcr1.GetName())
@@ -387,14 +321,14 @@ func TestUpdate(t *testing.T) {
 
 	// try update with wrong kind
 	tcr1.Meta.Kind = "WrongKind"
-	err = ds.Update(ctx, tcr1)
+	err = tenantDS.Update(ctx, tcr1)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "update of type:tenant failed, kind is set to:WrongKind but must be:Tenant")
 
 	// try update with wrong kind
 	tcr1.Meta.Kind = "Tenant"
 	tcr1.Meta.Apiversion = "v2"
-	err = ds.Update(ctx, tcr1)
+	err = tenantDS.Update(ctx, tcr1)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "update of type:tenant failed, apiversion must be set to:v1")
 
@@ -403,16 +337,20 @@ func TestUpdate(t *testing.T) {
 
 //nolint:unparam
 func checkHistoryCreated(ctx context.Context, t *testing.T, id string, name string, desc string) {
+	tenantDS, err := New(zaptest.NewLogger(t), db, &v1.Tenant{})
+	require.NoError(t, err)
 	var tgrhc v1.Tenant
-	err := ds.GetHistoryCreated(ctx, id, &tgrhc)
+	err = tenantDS.GetHistoryCreated(ctx, id, &tgrhc)
 	assert.NoError(t, err)
 	assert.Equal(t, name, tgrhc.Name)
 	assert.Equal(t, desc, tgrhc.GetDescription())
 }
 
 func checkHistory(ctx context.Context, t *testing.T, id string, tm time.Time, name string, desc string) {
+	tenantDS, err := New(zaptest.NewLogger(t), db, &v1.Tenant{})
+	require.NoError(t, err)
 	var tgrh v1.Tenant
-	err := ds.GetHistory(ctx, id, tm, &tgrh)
+	err = tenantDS.GetHistory(ctx, id, tm, &tgrh)
 	assert.NoError(t, err)
 	assert.Equal(t, name, tgrh.Name)
 	assert.Equal(t, desc, tgrh.GetDescription())
@@ -420,20 +358,14 @@ func checkHistory(ctx context.Context, t *testing.T, id string, tm time.Time, na
 
 func TestGet(t *testing.T) {
 	const t4 = "t4"
-	assert.NotNil(t, ds, "Datastore must not be nil")
+	tenantDS, err := New(zaptest.NewLogger(t), db, &v1.Tenant{})
+	require.NoError(t, err)
+	assert.NotNil(t, tenantDS, "Datastore must not be nil")
 	ctx := context.Background()
-	// unregistered type
-
-	ive := &invalidVersionedEntity{}
-	err := ds.Get(ctx, "", ive)
-	assert.Error(t, err)
-	assert.EqualError(t, err, "type:invalid is not registered")
-
 	// unknown id
-	var tgr1 v1.Tenant
-	err = ds.Get(ctx, "unknown-id", &tgr1)
+	_, err = tenantDS.Get(ctx, "unknown-id")
 	assert.Error(t, err)
-	assert.EqualError(t, err, "entity of type:tenant with id:unknown-id not found")
+	assert.EqualError(t, err, "tenant with id:unknown-id not found sql: no rows in result set")
 
 	// create a tenant
 	tcr1 := &v1.Tenant{
@@ -441,15 +373,14 @@ func TestGet(t *testing.T) {
 		Name:        "dtenant",
 		Description: "D Tenant",
 	}
-	err = ds.Create(ctx, tcr1)
+	err = tenantDS.Create(ctx, tcr1)
 	assert.NoError(t, err)
 	assert.Equal(t, t4, tcr1.GetMeta().GetId())
 	assert.Equal(t, "dtenant", tcr1.GetName())
 	assert.Equal(t, "D Tenant", tcr1.GetDescription())
 
 	// now get it
-	var tgr2 v1.Tenant
-	err = ds.Get(ctx, t4, &tgr2)
+	tgr2, err := tenantDS.Get(ctx, t4)
 	assert.NoError(t, err)
 	assert.Equal(t, t4, tgr2.GetMeta().GetId())
 	assert.Equal(t, "dtenant", tgr2.GetName())
@@ -458,27 +389,20 @@ func TestGet(t *testing.T) {
 
 func TestGetHistory(t *testing.T) {
 	const t5 = "t5"
-	assert.NotNil(t, ds, "Datastore must not be nil")
+	tenantDS, err := New(zaptest.NewLogger(t), db, &v1.Tenant{})
+	require.NoError(t, err)
+	assert.NotNil(t, tenantDS, "Datastore must not be nil")
 	ctx := context.Background()
 
 	tsNow := time.Date(2020, 4, 30, 18, 0, 0, 0, time.UTC)
 
-	ive := &invalidVersionedEntity{}
-	err := ds.GetHistory(ctx, "", tsNow, ive)
-	assert.Error(t, err)
-	assert.EqualError(t, err, "type:invalid is not registered")
-
-	err = ds.GetHistoryCreated(ctx, "", ive)
-	assert.Error(t, err)
-	assert.EqualError(t, err, "type:invalid is not registered")
-
 	// unknown id
 	var tgr1 v1.Tenant
-	err = ds.GetHistory(ctx, "unknown-id", tsNow, &tgr1)
+	err = tenantDS.GetHistory(ctx, "unknown-id", tsNow, &tgr1)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "entity of type:tenant with predicate:[map[id:unknown-id] map[created_at:2020-04-30 18:00:00 +0000 UTC]] not found")
 
-	err = ds.GetHistoryCreated(ctx, "unknown-id", &tgr1)
+	err = tenantDS.GetHistoryCreated(ctx, "unknown-id", &tgr1)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "entity of type:tenant with predicate:[map[id:unknown-id] map[op:C]] not found")
 
@@ -488,11 +412,11 @@ func TestGetHistory(t *testing.T) {
 	defer resetNow()
 
 	var tgrH v1.Tenant
-	err = ds.GetHistory(ctx, t5, createTS, &tgrH)
+	err = tenantDS.GetHistory(ctx, t5, createTS, &tgrH)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "entity of type:tenant with predicate:[map[id:t5] map[created_at:2020-04-30 18:00:00 +0000 UTC]] not found")
 
-	err = ds.GetHistoryCreated(ctx, t5, &tgrH)
+	err = tenantDS.GetHistoryCreated(ctx, t5, &tgrH)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "entity of type:tenant with predicate:[map[id:t5] map[op:C]] not found")
 
@@ -502,21 +426,20 @@ func TestGetHistory(t *testing.T) {
 		Name:        "dtenant",
 		Description: "D Tenant",
 	}
-	err = ds.Create(ctx, tcr1)
+	err = tenantDS.Create(ctx, tcr1)
 	assert.NoError(t, err)
 
 	checkHistoryCreated(ctx, t, t5, "dtenant", "D Tenant")
 	checkHistory(ctx, t, t5, createTS, "dtenant", "D Tenant")
 
-	var tcrU v1.Tenant
-	err = ds.Get(ctx, t5, &tcrU)
+	tcrU, err := tenantDS.Get(ctx, t5)
 	assert.NoError(t, err)
 	assert.Equal(t, createTS, convertToTime(tcrU.Meta.CreatedTime))
 
 	updateTS := time.Date(2020, 4, 30, 20, 0, 0, 0, time.UTC)
 	setNow(updateTS)
 	tcrU.Name = "dtenant updated"
-	err = ds.Update(ctx, &tcrU)
+	err = tenantDS.Update(ctx, tcrU)
 	assert.NoError(t, err)
 	assert.Equal(t, updateTS, convertToTime(tcrU.Meta.UpdatedTime))
 
@@ -526,7 +449,7 @@ func TestGetHistory(t *testing.T) {
 	update2TS := time.Date(2020, 4, 30, 21, 0, 0, 0, time.UTC)
 	setNow(update2TS)
 	tcrU.Name = "dtenant updated 2"
-	err = ds.Update(ctx, &tcrU)
+	err = tenantDS.Update(ctx, tcrU)
 	assert.NoError(t, err)
 	assert.Equal(t, update2TS, convertToTime(tcrU.Meta.UpdatedTime))
 
@@ -535,7 +458,7 @@ func TestGetHistory(t *testing.T) {
 
 	deletedTS := time.Date(2020, 4, 30, 22, 0, 0, 0, time.UTC)
 	setNow(deletedTS)
-	err = ds.Delete(ctx, tcr1)
+	err = tenantDS.Delete(ctx, tcr1.Meta.Id)
 	assert.NoError(t, err)
 
 	checkHistoryCreated(ctx, t, t5, "dtenant", "D Tenant")
@@ -543,7 +466,7 @@ func TestGetHistory(t *testing.T) {
 
 	// Check complete history
 	// before create
-	err = ds.GetHistory(ctx, t5, time.Date(2019, 1, 1, 8, 0, 0, 0, time.UTC), &tgrH)
+	err = tenantDS.GetHistory(ctx, t5, time.Date(2019, 1, 1, 8, 0, 0, 0, time.UTC), &tgrH)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "entity of type:tenant with predicate:[map[id:t5] map[created_at:2019-01-01 08:00:00 +0000 UTC]] not found")
 
@@ -557,25 +480,10 @@ func TestGetHistory(t *testing.T) {
 
 func TestFind(t *testing.T) {
 	const t6 = "t6"
-	assert.NotNil(t, ds, "Datastore must not be nil")
+	tenantDS, err := New(zaptest.NewLogger(t), db, &v1.Tenant{})
+	require.NoError(t, err)
+	assert.NotNil(t, tenantDS, "Datastore must not be nil")
 	ctx := context.Background()
-	// result not a slice
-	var te v1.Tenant
-	_, err := ds.Find(ctx, nil, nil, &te)
-	assert.Error(t, err)
-	assert.EqualError(t, err, "result argument must be a slice address")
-
-	// result is no versionedEntity
-	var res []string
-	_, err = ds.Find(ctx, nil, nil, &res)
-	assert.Error(t, err)
-	assert.EqualError(t, err, "result slice element type must implement VersionedJSONEntity-Interface")
-
-	// unregistered type
-	var ives = []invalidVersionedEntity{}
-	_, err = ds.Find(ctx, nil, nil, &ives)
-	assert.Error(t, err)
-	assert.EqualError(t, err, "type:invalid is not registered")
 
 	// create a tenant
 	tcr1 := &v1.Tenant{
@@ -583,17 +491,16 @@ func TestFind(t *testing.T) {
 		Name:        "ftenant",
 		Description: "F Tenant",
 	}
-	err = ds.Create(ctx, tcr1)
+	err = tenantDS.Create(ctx, tcr1)
 	assert.NoError(t, err)
 	assert.Equal(t, t6, tcr1.GetMeta().GetId())
 	assert.Equal(t, "ftenant", tcr1.GetName())
 	assert.Equal(t, "F Tenant", tcr1.GetDescription())
 
 	// now search it
-	var tfr []v1.Tenant
-	filter := make(map[string]interface{})
+	filter := make(map[string]any)
 	filter["id"] = t6
-	_, err = ds.Find(ctx, filter, nil, &tfr)
+	tfr, _, err := tenantDS.Find(ctx, filter, nil)
 	assert.NoError(t, err)
 	assert.NotNil(t, tfr)
 	assert.Len(t, tfr, 1)
@@ -605,61 +512,94 @@ func TestFind(t *testing.T) {
 			Name:        fmt.Sprintf("tenant-%d", i),
 			Description: fmt.Sprintf("Tenant %d", i),
 		}
-		err = ds.Create(ctx, tcr)
+		err = tenantDS.Create(ctx, tcr)
 		assert.NoError(t, err)
 	}
 	// find all
-	filter = make(map[string]interface{})
-	_, err = ds.Find(ctx, filter, nil, &tfr)
+	filter = make(map[string]any)
+	tfr, _, err = tenantDS.Find(ctx, filter, nil)
 	assert.NoError(t, err)
 	assert.NotNil(t, tfr)
 
 	// find one
-	var t9 []v1.Tenant
 	filter["id"] = "ftenant-9"
-	_, err = ds.Find(ctx, filter, nil, &t9)
+	t9, _, err := tenantDS.Find(ctx, filter, nil)
 	assert.NoError(t, err)
 	assert.NotNil(t, t9)
 	assert.Len(t, t9, 1)
 
 	// find one by name
-	var t8 []v1.Tenant
-	filter = make(map[string]interface{})
+	filter = make(map[string]any)
 	filter["tenant ->> 'name'"] = "tenant-8"
-	_, err = ds.Find(ctx, filter, nil, &t8)
+	t8, _, err := tenantDS.Find(ctx, filter, nil)
 	assert.NoError(t, err)
 	assert.NotNil(t, t8)
 	assert.Len(t, t8, 1)
 
 	// find one by description
-	var t4 []v1.Tenant
-	filter = make(map[string]interface{})
+	filter = make(map[string]any)
 	filter["tenant ->> 'description'"] = "Tenant 4"
-	_, err = ds.Find(ctx, filter, nil, &t4)
+	t4, _, err := tenantDS.Find(ctx, filter, nil)
 	assert.NoError(t, err)
 	assert.NotNil(t, t4)
 	assert.Len(t, t4, 1)
 
 }
 
+func TestFindWithPaging(t *testing.T) {
+	tenantDS, err := New(zaptest.NewLogger(t), db, &v1.Tenant{})
+	// prevent side effects
+	db.MustExec("DELETE from tenants")
+	require.NoError(t, err)
+	assert.NotNil(t, tenantDS, "Datastore must not be nil")
+	ctx := context.Background()
+
+	// create some tenants
+
+	count := 100
+	for i := 0; i < count; i++ {
+		tn := &v1.Tenant{
+			Meta: &v1.Meta{Id: fmt.Sprintf("t-%d", i)},
+			Name: "paging",
+		}
+		err := tenantDS.Create(ctx, tn)
+		assert.NoError(t, err)
+	}
+
+	// First find all with no paging
+	ts, nextpage, err := tenantDS.Find(ctx, nil, nil)
+	assert.NoError(t, err)
+	assert.Nil(t, nextpage)
+	assert.Len(t, ts, 100)
+
+	// Then find the first 60 results
+	ts, nextpage, err = tenantDS.Find(ctx, nil, &v1.Paging{Count: pointer.Pointer(uint64(60))})
+	assert.NoError(t, err)
+	assert.NotNil(t, nextpage)
+	assert.Equal(t, uint64(1), *nextpage)
+	assert.Len(t, ts, 60)
+
+	// At least the next 60, but only 40 left and no more pages
+	ts, nextpage, err = tenantDS.Find(ctx, nil, &v1.Paging{Page: nextpage, Count: pointer.Pointer(uint64(60))})
+	assert.NoError(t, err)
+	assert.Nil(t, nextpage)
+	assert.Len(t, ts, 40)
+}
+
 func TestDelete(t *testing.T) {
 	const t9 = "t9"
-	assert.NotNil(t, ds, "Datastore must not be nil")
+	tenantDS, err := New(zaptest.NewLogger(t), db, &v1.Tenant{})
+	require.NoError(t, err)
+	assert.NotNil(t, tenantDS, "Datastore must not be nil")
 	ctx := context.Background()
-	// unregistered type
-
-	ive := &invalidVersionedEntity{}
-	err := ds.Delete(ctx, ive)
-	assert.Error(t, err)
-	assert.EqualError(t, err, "type:invalid is not registered")
 
 	// unknown id
 	tdr1 := &v1.Tenant{
 		Meta: &v1.Meta{Id: "unknown-id"},
 	}
-	err = ds.Delete(ctx, tdr1)
+	err = tenantDS.Delete(ctx, tdr1.Meta.Id)
 	assert.Error(t, err)
-	assert.EqualError(t, err, "entity of type:tenant with id:unknown-id not found")
+	assert.EqualError(t, err, "tenant with id:unknown-id not found sql: no rows in result set")
 
 	// create a tenant
 	tcr1 := &v1.Tenant{
@@ -667,7 +607,7 @@ func TestDelete(t *testing.T) {
 		Name:        "etenant",
 		Description: "E Tenant",
 	}
-	err = ds.Create(ctx, tcr1)
+	err = tenantDS.Create(ctx, tcr1)
 	assert.NoError(t, err)
 	assert.Equal(t, t9, tcr1.GetMeta().GetId())
 	assert.Equal(t, "etenant", tcr1.GetName())
@@ -677,22 +617,23 @@ func TestDelete(t *testing.T) {
 	tdr2 := &v1.Tenant{
 		Meta: &v1.Meta{Id: t9},
 	}
-	err = ds.Delete(ctx, tdr2)
+	err = tenantDS.Delete(ctx, tdr2.Meta.Id)
 	assert.NoError(t, err)
 
-	var tgr v1.Tenant
-	err = ds.Get(ctx, t9, &tgr)
+	_, err = tenantDS.Get(ctx, t9)
 	assert.Error(t, err)
-	assert.EqualError(t, err, "entity of type:tenant with id:t9 not found")
+	assert.EqualError(t, err, "tenant with id:t9 not found sql: no rows in result set")
 
 	var tgrh v1.Tenant
-	err = ds.GetHistory(ctx, t9, time.Now(), &tgrh)
+	err = tenantDS.GetHistory(ctx, t9, time.Now(), &tgrh)
 	assert.NoError(t, err)
 	assert.Equal(t, "etenant", tgrh.Name)
 }
 
 func TestAnnotationsAndLabels(t *testing.T) {
-	assert.NotNil(t, ds, "Datastore must not be nil")
+	tenantDS, err := New(zaptest.NewLogger(t), db, &v1.Tenant{})
+	require.NoError(t, err)
+	assert.NotNil(t, tenantDS, "Datastore must not be nil")
 	ctx := context.Background()
 	tcr := &v1.Tenant{
 		Meta: &v1.Meta{
@@ -709,7 +650,7 @@ func TestAnnotationsAndLabels(t *testing.T) {
 		Description: "A very important Tenant",
 	}
 
-	err := ds.Create(ctx, tcr)
+	err = tenantDS.Create(ctx, tcr)
 	assert.NoError(t, err)
 	assert.NotNil(t, tcr)
 	assert.Equal(t, int64(0), tcr.Meta.Version)
@@ -720,23 +661,21 @@ func TestAnnotationsAndLabels(t *testing.T) {
 	assert.Equal(t, []string{"color=red"}, tcr.GetMeta().GetLabels())
 
 	// get from db
-	tget := &v1.Tenant{}
-	err = ds.Get(ctx, tcr.Meta.Id, tget)
+	tget, err := tenantDS.Get(ctx, tcr.Meta.Id)
 	require.NoError(t, err)
 	require.Equal(t, tcr.Meta.Id, tget.Meta.Id)
 	assert.Equal(t, int64(0), tget.Meta.Version)
 
 	// update instance
 	tcr.Name = "updated name"
-	err = ds.Update(ctx, tcr)
+	err = tenantDS.Update(ctx, tcr)
 	assert.NoError(t, err)
 	assert.NotNil(t, tcr)
 	// incremented version after update
 	assert.Equal(t, int64(1), tcr.Meta.Version)
 
 	// re-read from db
-	var tgr v1.Tenant
-	err = ds.Get(ctx, tcr.Meta.GetId(), &tgr)
+	tgr, err := tenantDS.Get(ctx, tcr.Meta.GetId())
 	assert.NoError(t, err)
 	// version is incremented
 	assert.Equal(t, int64(1), tgr.Meta.Version)
@@ -745,7 +684,7 @@ func TestAnnotationsAndLabels(t *testing.T) {
 
 	// try to update older version --> optimistic lock error
 	tget.Name = "updated older entity"
-	err = ds.Update(ctx, tget)
+	err = tenantDS.Update(ctx, tget)
 	require.Equal(t, err,
 		NewOptimisticLockError(
 			fmt.Sprintf("optimistic lock error updating tenant with id %s, existing version 1 mismatches entity version 0", tget.GetMeta().Id),
@@ -758,7 +697,7 @@ func TestAnnotationsAndLabels(t *testing.T) {
 	ls := []string{"color=red", "size=xlarge"}
 	tcr.GetMeta().SetAnnotations(as)
 	tcr.GetMeta().SetLabels(ls)
-	err = ds.Update(ctx, tcr)
+	err = tenantDS.Update(ctx, tcr)
 	assert.NoError(t, err)
 	assert.NotNil(t, tcr)
 	assert.Equal(t, map[string]string{"key1": "value3", "key2": "value2"}, tcr.GetMeta().GetAnnotations())
@@ -779,4 +718,54 @@ func setNow(t time.Time) {
 // resetNow resets the overriden Now to time.Now
 func resetNow() {
 	Now = time.Now
+}
+
+func createPostgresConnection() (*sqlx.DB, error) {
+
+	ctx := context.Background()
+	req := testcontainers.ContainerRequest{
+		Image:        "postgres:15-alpine",
+		ExposedPorts: []string{"5432/tcp"},
+		Env:          map[string]string{"POSTGRES_PASSWORD": "password"},
+		// TODO: should work, but dont, hence using the loop below to check pg is up an ready for connections.
+		// WaitingFor:   wait.ForLog("database system is ready to accept connections"),
+		// WaitingFor: wait.ForListeningPort("5432/tcp"),
+	}
+
+	postgres, err := testcontainers.GenericContainer(ctx, testcontainers.GenericContainerRequest{
+		ContainerRequest: req,
+		Started:          true,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	ip, err := postgres.Host(ctx)
+	if err != nil {
+		return nil, err
+	}
+	port, err := postgres.MappedPort(ctx, "5432/tcp")
+	if err != nil {
+		return nil, err
+	}
+
+	log := zap.NewNop()
+	var db *sqlx.DB
+	for {
+		var err error
+		ves := []Entity{
+			&v1.Project{},
+			&v1.Tenant{},
+		}
+		db, err = NewPostgresDB(log, ip, port.Port(), "postgres", "password", "postgres", "disable", ves...)
+		if err != nil {
+			time.Sleep(100 * time.Millisecond)
+			continue
+		}
+		err = db.Ping()
+		if err == nil {
+			break
+		}
+	}
+	return db, nil
 }
