@@ -6,6 +6,7 @@ import (
 	v1 "github.com/metal-stack/masterdata-api/api/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
+	"go.uber.org/zap/zaptest"
 	"google.golang.org/protobuf/types/known/wrapperspb"
 
 	"testing"
@@ -14,8 +15,13 @@ import (
 )
 
 func TestCreateProject(t *testing.T) {
-	storageMock := &mocks.Storage{}
-	ts := NewProjectService(storageMock, log)
+	storageMock := &mocks.Storage[*v1.Project]{}
+	tenantStorageMock := &mocks.Storage[*v1.Tenant]{}
+	ts := &projectService{
+		projectStore: storageMock,
+		tenantStore:  tenantStorageMock,
+		log:          zaptest.NewLogger(t),
+	}
 	ctx := context.Background()
 
 	t1 := &v1.Tenant{}
@@ -35,7 +41,7 @@ func TestCreateProject(t *testing.T) {
 	tcr := &v1.ProjectCreateRequest{
 		Project: p1,
 	}
-	storageMock.On("Get", ctx, p1.GetTenantId(), t1).Return(nil)
+	tenantStorageMock.On("Get", ctx, p1.GetTenantId()).Return(t1, nil)
 	storageMock.On("Create", ctx, p1).Return(nil)
 	resp, err := ts.Create(ctx, tcr)
 	assert.NoError(t, err)
@@ -45,8 +51,13 @@ func TestCreateProject(t *testing.T) {
 }
 
 func TestCreateProjectWithQuotaCheck(t *testing.T) {
-	storageMock := &mocks.Storage{}
-	ts := NewProjectService(storageMock, log)
+	storageMock := &mocks.Storage[*v1.Project]{}
+	tenantStorageMock := &mocks.Storage[*v1.Tenant]{}
+	ts := &projectService{
+		projectStore: storageMock,
+		tenantStore:  tenantStorageMock,
+		log:          zaptest.NewLogger(t),
+	}
 	ctx := context.Background()
 
 	t1 := &v1.Tenant{
@@ -64,15 +75,12 @@ func TestCreateProjectWithQuotaCheck(t *testing.T) {
 	tcr := &v1.ProjectCreateRequest{
 		Project: p1,
 	}
-	filter := make(map[string]interface{})
+	filter := make(map[string]any)
 	filter["project ->> 'tenant_id'"] = p1.TenantId
-	var projects []v1.Project
+	var projects []*v1.Project
 	// see: https://github.com/stretchr/testify/blob/master/mock/mock.go#L149-L162
-	storageMock.On("Get", ctx, p1.GetTenantId(), mock.AnythingOfType("*v1.Tenant")).Return(nil).Run(func(args mock.Arguments) {
-		arg := args.Get(2).(*v1.Tenant)
-		arg.Quotas = t1.GetQuotas()
-	})
-	storageMock.On("Find", ctx, filter, mock.AnythingOfType("*v1.Paging"), &projects).Return(nil, nil)
+	tenantStorageMock.On("Get", ctx, p1.GetTenantId()).Return(t1, nil)
+	storageMock.On("Find", ctx, filter, mock.AnythingOfType("*v1.Paging")).Return(projects, nil, nil)
 	storageMock.On("Create", ctx, p1).Return(nil)
 	resp, err := ts.Create(ctx, tcr)
 	assert.NoError(t, err)
@@ -82,8 +90,13 @@ func TestCreateProjectWithQuotaCheck(t *testing.T) {
 }
 
 func TestUpdateProject(t *testing.T) {
-	storageMock := &mocks.Storage{}
-	ts := NewProjectService(storageMock, log)
+	storageMock := &mocks.Storage[*v1.Project]{}
+	tenantStorageMock := &mocks.Storage[*v1.Tenant]{}
+	ts := &projectService{
+		projectStore: storageMock,
+		tenantStore:  tenantStorageMock,
+		log:          zaptest.NewLogger(t),
+	}
 	ctx := context.Background()
 
 	t1 := &v1.Project{
@@ -108,8 +121,13 @@ func TestUpdateProject(t *testing.T) {
 }
 
 func TestDeleteProject(t *testing.T) {
-	storageMock := &mocks.Storage{}
-	ts := NewProjectService(storageMock, log)
+	storageMock := &mocks.Storage[*v1.Project]{}
+	tenantStorageMock := &mocks.Storage[*v1.Tenant]{}
+	ts := &projectService{
+		projectStore: storageMock,
+		tenantStore:  tenantStorageMock,
+		log:          zaptest.NewLogger(t),
+	}
 	ctx := context.Background()
 	t3 := &v1.Project{
 		Meta: &v1.Meta{Id: "p3"},
@@ -118,7 +136,7 @@ func TestDeleteProject(t *testing.T) {
 		Id: "p3",
 	}
 
-	storageMock.On("Delete", ctx, t3).Return(nil)
+	storageMock.On("Delete", ctx, t3.Meta.Id).Return(nil)
 	resp, err := ts.Delete(ctx, tdr)
 	assert.NoError(t, err)
 	assert.NotNil(t, resp)
@@ -126,75 +144,96 @@ func TestDeleteProject(t *testing.T) {
 	assert.Equal(t, tdr.Id, resp.GetProject().GetMeta().GetId())
 }
 
-// FIXME reenable
-// func TestGetProject(t *testing.T) {
-// 	storageMock := &mocks.Storage{}
-// 	ts := NewProjectService(storageMock, log)
-// 	ctx := context.Background()
-// 	t4 := &v1.Project{}
-// 	tgr := &v1.ProjectGetRequest{
-// 		Id: "p4",
-// 	}
+func TestGetProject(t *testing.T) {
+	storageMock := &mocks.Storage[*v1.Project]{}
+	tenantStorageMock := &mocks.Storage[*v1.Tenant]{}
+	ts := &projectService{
+		projectStore: storageMock,
+		tenantStore:  tenantStorageMock,
+		log:          zaptest.NewLogger(t),
+	}
+	ctx := context.Background()
+	t4 := &v1.Project{
+		Meta: &v1.Meta{Id: "p4"},
+	}
+	tgr := &v1.ProjectGetRequest{
+		Id: "p4",
+	}
 
-// 	storageMock.On("Get", ctx, "p4", t4).Return(nil)
-// 	resp, err := ts.Get(ctx, tgr)
-// 	assert.NoError(t, err)
-// 	assert.NotNil(t, resp)
-// 	assert.NotNil(t, resp.GetProject())
-// 	assert.Equal(t, tgr.Id, resp.GetProject().GetMeta().GetId())
-// }
+	storageMock.On("Get", ctx, "p4").Return(t4, nil)
+	resp, err := ts.Get(ctx, tgr)
+	assert.NoError(t, err)
+	assert.NotNil(t, resp)
+	assert.NotNil(t, resp.GetProject())
+	assert.Equal(t, tgr.Id, resp.GetProject().GetMeta().GetId())
+}
 
 func TestFindProjectByID(t *testing.T) {
-	storageMock := &mocks.Storage{}
-	ts := NewProjectService(storageMock, log)
+	storageMock := &mocks.Storage[*v1.Project]{}
+	tenantStorageMock := &mocks.Storage[*v1.Tenant]{}
+	ts := &projectService{
+		projectStore: storageMock,
+		tenantStore:  tenantStorageMock,
+		log:          zaptest.NewLogger(t),
+	}
 	ctx := context.Background()
-	var t5s []v1.Project
+	var t5s []*v1.Project
 	// filter by id
-	f1 := make(map[string]interface{})
+	f1 := make(map[string]any)
 	tfr := &v1.ProjectFindRequest{
 		Id: &wrapperspb.StringValue{Value: "p5"},
 	}
 
 	f1["id"] = "p5"
-	storageMock.On("Find", ctx, f1, mock.AnythingOfType("*v1.Paging"), &t5s).Return(nil, nil)
+	storageMock.On("Find", ctx, f1, mock.AnythingOfType("*v1.Paging")).Return(t5s, nil, nil)
 	resp, err := ts.Find(ctx, tfr)
 	assert.NoError(t, err)
 	assert.NotNil(t, resp)
 }
 
 func TestFindProjectByName(t *testing.T) {
-	storageMock := &mocks.Storage{}
-	ts := NewProjectService(storageMock, log)
+	storageMock := &mocks.Storage[*v1.Project]{}
+	tenantStorageMock := &mocks.Storage[*v1.Tenant]{}
+	ts := &projectService{
+		projectStore: storageMock,
+		tenantStore:  tenantStorageMock,
+		log:          zaptest.NewLogger(t),
+	}
 	ctx := context.Background()
 
 	// filter by name
-	var t6s []v1.Project
+	var t6s []*v1.Project
 	tfr := &v1.ProjectFindRequest{
 		Name: &wrapperspb.StringValue{Value: "Sixth"},
 	}
 
-	f2 := make(map[string]interface{})
+	f2 := make(map[string]any)
 	f2["project ->> 'name'"] = "Sixth"
-	storageMock.On("Find", ctx, f2, mock.AnythingOfType("*v1.Paging"), &t6s).Return(nil, nil)
+	storageMock.On("Find", ctx, f2, mock.AnythingOfType("*v1.Paging")).Return(t6s, nil, nil)
 	resp, err := ts.Find(ctx, tfr)
 	assert.NoError(t, err)
 	assert.NotNil(t, resp)
 }
 
 func TestFindProjectByTenant(t *testing.T) {
-	storageMock := &mocks.Storage{}
-	ts := NewProjectService(storageMock, log)
+	storageMock := &mocks.Storage[*v1.Project]{}
+	tenantStorageMock := &mocks.Storage[*v1.Tenant]{}
+	ts := &projectService{
+		projectStore: storageMock,
+		tenantStore:  tenantStorageMock,
+		log:          zaptest.NewLogger(t),
+	}
 	ctx := context.Background()
 
 	// filter by name
-	var t6s []v1.Project
+	var t6s []*v1.Project
 	tfr := &v1.ProjectFindRequest{
 		TenantId: &wrapperspb.StringValue{Value: "p1"},
 	}
 
-	f2 := make(map[string]interface{})
+	f2 := make(map[string]any)
 	f2["project ->> 'tenant_id'"] = "p1"
-	storageMock.On("Find", ctx, f2, mock.AnythingOfType("*v1.Paging"), &t6s).Return(nil, nil)
+	storageMock.On("Find", ctx, f2, mock.AnythingOfType("*v1.Paging")).Return(t6s, nil, nil)
 	resp, err := ts.Find(ctx, tfr)
 	assert.NoError(t, err)
 	assert.NotNil(t, resp)
