@@ -4,47 +4,51 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/jmoiron/sqlx"
 	v1 "github.com/metal-stack/masterdata-api/api/v1"
 	"github.com/metal-stack/masterdata-api/pkg/datastore"
 	"go.uber.org/zap"
 )
 
-type TenantService struct {
-	Storage datastore.Storage
-	log     *zap.Logger
+type tenantService struct {
+	tenantStore datastore.Storage[*v1.Tenant]
+	log         *zap.Logger
 }
 
-func NewTenantService(s datastore.Storage, l *zap.Logger) *TenantService {
-	return &TenantService{
-		Storage: NewStorageStatusWrapper(s),
-		log:     l,
+func NewTenantService(db *sqlx.DB, l *zap.Logger) (*tenantService, error) {
+	ts, err := datastore.New(l, db, &v1.Tenant{})
+	if err != nil {
+		return nil, err
 	}
+	return &tenantService{
+		tenantStore: NewStorageStatusWrapper(ts),
+		log:         l,
+	}, nil
 }
 
-func (s *TenantService) Create(ctx context.Context, req *v1.TenantCreateRequest) (*v1.TenantResponse, error) {
+func (s *tenantService) Create(ctx context.Context, req *v1.TenantCreateRequest) (*v1.TenantResponse, error) {
 	tenant := req.Tenant
 	// allow create without sending Meta
 	if tenant.Meta == nil {
 		tenant.Meta = &v1.Meta{}
 	}
-	err := s.Storage.Create(ctx, tenant)
+	err := s.tenantStore.Create(ctx, tenant)
 	return tenant.NewTenantResponse(), err
 }
-func (s *TenantService) Update(ctx context.Context, req *v1.TenantUpdateRequest) (*v1.TenantResponse, error) {
+func (s *tenantService) Update(ctx context.Context, req *v1.TenantUpdateRequest) (*v1.TenantResponse, error) {
 	tenant := req.Tenant
-	err := s.Storage.Update(ctx, tenant)
+	err := s.tenantStore.Update(ctx, tenant)
 	return tenant.NewTenantResponse(), err
 }
 
-func (s *TenantService) Delete(ctx context.Context, req *v1.TenantDeleteRequest) (*v1.TenantResponse, error) {
+func (s *tenantService) Delete(ctx context.Context, req *v1.TenantDeleteRequest) (*v1.TenantResponse, error) {
 	tenant := req.NewTenant()
-	err := s.Storage.Delete(ctx, tenant)
+	err := s.tenantStore.Delete(ctx, tenant.Meta.Id)
 	return tenant.NewTenantResponse(), err
 }
 
-func (s *TenantService) Get(ctx context.Context, req *v1.TenantGetRequest) (*v1.TenantResponse, error) {
-	tenant := &v1.Tenant{}
-	err := s.Storage.Get(ctx, req.Id, tenant)
+func (s *tenantService) Get(ctx context.Context, req *v1.TenantGetRequest) (*v1.TenantResponse, error) {
+	tenant, err := s.tenantStore.Get(ctx, req.Id)
 	if err != nil {
 		return nil, err
 	}
@@ -53,11 +57,11 @@ func (s *TenantService) Get(ctx context.Context, req *v1.TenantGetRequest) (*v1.
 	return tenant.NewTenantResponse(), nil
 }
 
-func (s *TenantService) GetHistory(ctx context.Context, req *v1.TenantGetHistoryRequest) (*v1.TenantResponse, error) {
+func (s *tenantService) GetHistory(ctx context.Context, req *v1.TenantGetHistoryRequest) (*v1.TenantResponse, error) {
 	tenant := &v1.Tenant{}
 	at := req.At.AsTime()
 	s.log.Info("getHistory", zap.String("id", req.Id), zap.Time("at", at))
-	err := s.Storage.GetHistory(ctx, req.Id, at, tenant)
+	err := s.tenantStore.GetHistory(ctx, req.Id, at, tenant)
 	if err != nil {
 		return nil, err
 	}
@@ -66,9 +70,8 @@ func (s *TenantService) GetHistory(ctx context.Context, req *v1.TenantGetHistory
 	return tenant.NewTenantResponse(), nil
 }
 
-func (s *TenantService) Find(ctx context.Context, req *v1.TenantFindRequest) (*v1.TenantListResponse, error) {
-	var res []v1.Tenant
-	filter := make(map[string]interface{})
+func (s *tenantService) Find(ctx context.Context, req *v1.TenantFindRequest) (*v1.TenantListResponse, error) {
+	filter := make(map[string]any)
 	if req.Id != nil {
 		filter["id"] = req.GetId().GetValue()
 	}
@@ -80,13 +83,13 @@ func (s *TenantService) Find(ctx context.Context, req *v1.TenantFindRequest) (*v
 		f := fmt.Sprintf("tenant -> 'meta' -> 'annotations' ->> '%s'", key)
 		filter[f] = value
 	}
-	nextPage, err := s.Storage.Find(ctx, filter, req.Paging, &res)
+	res, nextPage, err := s.tenantStore.Find(ctx, filter, req.Paging)
 	if err != nil {
 		return nil, err
 	}
 	resp := new(v1.TenantListResponse)
 	for i := range res {
-		t := &res[i]
+		t := res[i]
 		resp.Tenants = append(resp.Tenants, t)
 	}
 	resp.NextPage = nextPage
