@@ -1,4 +1,4 @@
-// Code generated by mockery v2.14.0. DO NOT EDIT.
+// Code generated by mockery v2.24.0. DO NOT EDIT.
 
 package mocks
 
@@ -29,6 +29,10 @@ func (_m *ProjectServiceClient) Create(ctx context.Context, in *v1.ProjectCreate
 	ret := _m.Called(_ca...)
 
 	var r0 *v1.ProjectResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectCreateRequest, ...grpc.CallOption) (*v1.ProjectResponse, error)); ok {
+		return rf(ctx, in, opts...)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectCreateRequest, ...grpc.CallOption) *v1.ProjectResponse); ok {
 		r0 = rf(ctx, in, opts...)
 	} else {
@@ -37,7 +41,6 @@ func (_m *ProjectServiceClient) Create(ctx context.Context, in *v1.ProjectCreate
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.ProjectCreateRequest, ...grpc.CallOption) error); ok {
 		r1 = rf(ctx, in, opts...)
 	} else {
@@ -59,6 +62,10 @@ func (_m *ProjectServiceClient) Delete(ctx context.Context, in *v1.ProjectDelete
 	ret := _m.Called(_ca...)
 
 	var r0 *v1.ProjectResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectDeleteRequest, ...grpc.CallOption) (*v1.ProjectResponse, error)); ok {
+		return rf(ctx, in, opts...)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectDeleteRequest, ...grpc.CallOption) *v1.ProjectResponse); ok {
 		r0 = rf(ctx, in, opts...)
 	} else {
@@ -67,7 +74,6 @@ func (_m *ProjectServiceClient) Delete(ctx context.Context, in *v1.ProjectDelete
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.ProjectDeleteRequest, ...grpc.CallOption) error); ok {
 		r1 = rf(ctx, in, opts...)
 	} else {
@@ -89,6 +95,10 @@ func (_m *ProjectServiceClient) Find(ctx context.Context, in *v1.ProjectFindRequ
 	ret := _m.Called(_ca...)
 
 	var r0 *v1.ProjectListResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectFindRequest, ...grpc.CallOption) (*v1.ProjectListResponse, error)); ok {
+		return rf(ctx, in, opts...)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectFindRequest, ...grpc.CallOption) *v1.ProjectListResponse); ok {
 		r0 = rf(ctx, in, opts...)
 	} else {
@@ -97,7 +107,6 @@ func (_m *ProjectServiceClient) Find(ctx context.Context, in *v1.ProjectFindRequ
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.ProjectFindRequest, ...grpc.CallOption) error); ok {
 		r1 = rf(ctx, in, opts...)
 	} else {
@@ -119,6 +128,10 @@ func (_m *ProjectServiceClient) Get(ctx context.Context, in *v1.ProjectGetReques
 	ret := _m.Called(_ca...)
 
 	var r0 *v1.ProjectResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectGetRequest, ...grpc.CallOption) (*v1.ProjectResponse, error)); ok {
+		return rf(ctx, in, opts...)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectGetRequest, ...grpc.CallOption) *v1.ProjectResponse); ok {
 		r0 = rf(ctx, in, opts...)
 	} else {
@@ -127,7 +140,6 @@ func (_m *ProjectServiceClient) Get(ctx context.Context, in *v1.ProjectGetReques
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.ProjectGetRequest, ...grpc.CallOption) error); ok {
 		r1 = rf(ctx, in, opts...)
 	} else {
@@ -149,6 +161,10 @@ func (_m *ProjectServiceClient) GetHistory(ctx context.Context, in *v1.ProjectGe
 	ret := _m.Called(_ca...)
 
 	var r0 *v1.ProjectResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectGetHistoryRequest, ...grpc.CallOption) (*v1.ProjectResponse, error)); ok {
+		return rf(ctx, in, opts...)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectGetHistoryRequest, ...grpc.CallOption) *v1.ProjectResponse); ok {
 		r0 = rf(ctx, in, opts...)
 	} else {
@@ -157,7 +173,6 @@ func (_m *ProjectServiceClient) GetHistory(ctx context.Context, in *v1.ProjectGe
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.ProjectGetHistoryRequest, ...grpc.CallOption) error); ok {
 		r1 = rf(ctx, in, opts...)
 	} else {
@@ -179,6 +194,10 @@ func (_m *ProjectServiceClient) Update(ctx context.Context, in *v1.ProjectUpdate
 	ret := _m.Called(_ca...)
 
 	var r0 *v1.ProjectResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectUpdateRequest, ...grpc.CallOption) (*v1.ProjectResponse, error)); ok {
+		return rf(ctx, in, opts...)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectUpdateRequest, ...grpc.CallOption) *v1.ProjectResponse); ok {
 		r0 = rf(ctx, in, opts...)
 	} else {
@@ -187,7 +206,6 @@ func (_m *ProjectServiceClient) Update(ctx context.Context, in *v1.ProjectUpdate
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.ProjectUpdateRequest, ...grpc.CallOption) error); ok {
 		r1 = rf(ctx, in, opts...)
 	} else {
