@@ -1,4 +1,4 @@
-// Code generated by mockery v2.14.0. DO NOT EDIT.
+// Code generated by mockery v2.24.0. DO NOT EDIT.
 
 package mocks
 
@@ -19,6 +19,10 @@ func (_m *ProjectServiceServer) Create(_a0 context.Context, _a1 *v1.ProjectCreat
 	ret := _m.Called(_a0, _a1)
 
 	var r0 *v1.ProjectResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectCreateRequest) (*v1.ProjectResponse, error)); ok {
+		return rf(_a0, _a1)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectCreateRequest) *v1.ProjectResponse); ok {
 		r0 = rf(_a0, _a1)
 	} else {
@@ -27,7 +31,6 @@ func (_m *ProjectServiceServer) Create(_a0 context.Context, _a1 *v1.ProjectCreat
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.ProjectCreateRequest) error); ok {
 		r1 = rf(_a0, _a1)
 	} else {
@@ -42,6 +45,10 @@ func (_m *ProjectServiceServer) Delete(_a0 context.Context, _a1 *v1.ProjectDelet
 	ret := _m.Called(_a0, _a1)
 
 	var r0 *v1.ProjectResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectDeleteRequest) (*v1.ProjectResponse, error)); ok {
+		return rf(_a0, _a1)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectDeleteRequest) *v1.ProjectResponse); ok {
 		r0 = rf(_a0, _a1)
 	} else {
@@ -50,7 +57,6 @@ func (_m *ProjectServiceServer) Delete(_a0 context.Context, _a1 *v1.ProjectDelet
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.ProjectDeleteRequest) error); ok {
 		r1 = rf(_a0, _a1)
 	} else {
@@ -65,6 +71,10 @@ func (_m *ProjectServiceServer) Find(_a0 context.Context, _a1 *v1.ProjectFindReq
 	ret := _m.Called(_a0, _a1)
 
 	var r0 *v1.ProjectListResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectFindRequest) (*v1.ProjectListResponse, error)); ok {
+		return rf(_a0, _a1)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectFindRequest) *v1.ProjectListResponse); ok {
 		r0 = rf(_a0, _a1)
 	} else {
@@ -73,7 +83,6 @@ func (_m *ProjectServiceServer) Find(_a0 context.Context, _a1 *v1.ProjectFindReq
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.ProjectFindRequest) error); ok {
 		r1 = rf(_a0, _a1)
 	} else {
@@ -88,6 +97,10 @@ func (_m *ProjectServiceServer) Get(_a0 context.Context, _a1 *v1.ProjectGetReque
 	ret := _m.Called(_a0, _a1)
 
 	var r0 *v1.ProjectResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectGetRequest) (*v1.ProjectResponse, error)); ok {
+		return rf(_a0, _a1)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectGetRequest) *v1.ProjectResponse); ok {
 		r0 = rf(_a0, _a1)
 	} else {
@@ -96,7 +109,6 @@ func (_m *ProjectServiceServer) Get(_a0 context.Context, _a1 *v1.ProjectGetReque
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.ProjectGetRequest) error); ok {
 		r1 = rf(_a0, _a1)
 	} else {
@@ -111,6 +123,10 @@ func (_m *ProjectServiceServer) GetHistory(_a0 context.Context, _a1 *v1.ProjectG
 	ret := _m.Called(_a0, _a1)
 
 	var r0 *v1.ProjectResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectGetHistoryRequest) (*v1.ProjectResponse, error)); ok {
+		return rf(_a0, _a1)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectGetHistoryRequest) *v1.ProjectResponse); ok {
 		r0 = rf(_a0, _a1)
 	} else {
@@ -119,7 +135,6 @@ func (_m *ProjectServiceServer) GetHistory(_a0 context.Context, _a1 *v1.ProjectG
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.ProjectGetHistoryRequest) error); ok {
 		r1 = rf(_a0, _a1)
 	} else {
@@ -134,6 +149,10 @@ func (_m *ProjectServiceServer) Update(_a0 context.Context, _a1 *v1.ProjectUpdat
 	ret := _m.Called(_a0, _a1)
 
 	var r0 *v1.ProjectResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectUpdateRequest) (*v1.ProjectResponse, error)); ok {
+		return rf(_a0, _a1)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ProjectUpdateRequest) *v1.ProjectResponse); ok {
 		r0 = rf(_a0, _a1)
 	} else {
@@ -142,7 +161,6 @@ func (_m *ProjectServiceServer) Update(_a0 context.Context, _a1 *v1.ProjectUpdat
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.ProjectUpdateRequest) error); ok {
 		r1 = rf(_a0, _a1)
 	} else {
