@@ -1,4 +1,4 @@
-// Code generated by mockery v2.14.0. DO NOT EDIT.
+// Code generated by mockery v2.24.0. DO NOT EDIT.
 
 package mocks
 
@@ -29,6 +29,10 @@ func (_m *TenantServiceClient) Create(ctx context.Context, in *v1.TenantCreateRe
 	ret := _m.Called(_ca...)
 
 	var r0 *v1.TenantResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantCreateRequest, ...grpc.CallOption) (*v1.TenantResponse, error)); ok {
+		return rf(ctx, in, opts...)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantCreateRequest, ...grpc.CallOption) *v1.TenantResponse); ok {
 		r0 = rf(ctx, in, opts...)
 	} else {
@@ -37,7 +41,6 @@ func (_m *TenantServiceClient) Create(ctx context.Context, in *v1.TenantCreateRe
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.TenantCreateRequest, ...grpc.CallOption) error); ok {
 		r1 = rf(ctx, in, opts...)
 	} else {
@@ -59,6 +62,10 @@ func (_m *TenantServiceClient) Delete(ctx context.Context, in *v1.TenantDeleteRe
 	ret := _m.Called(_ca...)
 
 	var r0 *v1.TenantResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantDeleteRequest, ...grpc.CallOption) (*v1.TenantResponse, error)); ok {
+		return rf(ctx, in, opts...)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantDeleteRequest, ...grpc.CallOption) *v1.TenantResponse); ok {
 		r0 = rf(ctx, in, opts...)
 	} else {
@@ -67,7 +74,6 @@ func (_m *TenantServiceClient) Delete(ctx context.Context, in *v1.TenantDeleteRe
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.TenantDeleteRequest, ...grpc.CallOption) error); ok {
 		r1 = rf(ctx, in, opts...)
 	} else {
@@ -89,6 +95,10 @@ func (_m *TenantServiceClient) Find(ctx context.Context, in *v1.TenantFindReques
 	ret := _m.Called(_ca...)
 
 	var r0 *v1.TenantListResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantFindRequest, ...grpc.CallOption) (*v1.TenantListResponse, error)); ok {
+		return rf(ctx, in, opts...)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantFindRequest, ...grpc.CallOption) *v1.TenantListResponse); ok {
 		r0 = rf(ctx, in, opts...)
 	} else {
@@ -97,7 +107,6 @@ func (_m *TenantServiceClient) Find(ctx context.Context, in *v1.TenantFindReques
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.TenantFindRequest, ...grpc.CallOption) error); ok {
 		r1 = rf(ctx, in, opts...)
 	} else {
@@ -119,6 +128,10 @@ func (_m *TenantServiceClient) Get(ctx context.Context, in *v1.TenantGetRequest,
 	ret := _m.Called(_ca...)
 
 	var r0 *v1.TenantResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantGetRequest, ...grpc.CallOption) (*v1.TenantResponse, error)); ok {
+		return rf(ctx, in, opts...)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantGetRequest, ...grpc.CallOption) *v1.TenantResponse); ok {
 		r0 = rf(ctx, in, opts...)
 	} else {
@@ -127,7 +140,6 @@ func (_m *TenantServiceClient) Get(ctx context.Context, in *v1.TenantGetRequest,
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.TenantGetRequest, ...grpc.CallOption) error); ok {
 		r1 = rf(ctx, in, opts...)
 	} else {
@@ -149,6 +161,10 @@ func (_m *TenantServiceClient) GetHistory(ctx context.Context, in *v1.TenantGetH
 	ret := _m.Called(_ca...)
 
 	var r0 *v1.TenantResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantGetHistoryRequest, ...grpc.CallOption) (*v1.TenantResponse, error)); ok {
+		return rf(ctx, in, opts...)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantGetHistoryRequest, ...grpc.CallOption) *v1.TenantResponse); ok {
 		r0 = rf(ctx, in, opts...)
 	} else {
@@ -157,7 +173,6 @@ func (_m *TenantServiceClient) GetHistory(ctx context.Context, in *v1.TenantGetH
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.TenantGetHistoryRequest, ...grpc.CallOption) error); ok {
 		r1 = rf(ctx, in, opts...)
 	} else {
@@ -179,6 +194,10 @@ func (_m *TenantServiceClient) Update(ctx context.Context, in *v1.TenantUpdateRe
 	ret := _m.Called(_ca...)
 
 	var r0 *v1.TenantResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantUpdateRequest, ...grpc.CallOption) (*v1.TenantResponse, error)); ok {
+		return rf(ctx, in, opts...)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantUpdateRequest, ...grpc.CallOption) *v1.TenantResponse); ok {
 		r0 = rf(ctx, in, opts...)
 	} else {
@@ -187,7 +206,6 @@ func (_m *TenantServiceClient) Update(ctx context.Context, in *v1.TenantUpdateRe
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.TenantUpdateRequest, ...grpc.CallOption) error); ok {
 		r1 = rf(ctx, in, opts...)
 	} else {
