@@ -1,4 +1,4 @@
-// Code generated by mockery v2.14.0. DO NOT EDIT.
+// Code generated by mockery v2.24.0. DO NOT EDIT.
 
 package mocks
 
@@ -19,6 +19,10 @@ func (_m *TenantServiceServer) Create(_a0 context.Context, _a1 *v1.TenantCreateR
 	ret := _m.Called(_a0, _a1)
 
 	var r0 *v1.TenantResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantCreateRequest) (*v1.TenantResponse, error)); ok {
+		return rf(_a0, _a1)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantCreateRequest) *v1.TenantResponse); ok {
 		r0 = rf(_a0, _a1)
 	} else {
@@ -27,7 +31,6 @@ func (_m *TenantServiceServer) Create(_a0 context.Context, _a1 *v1.TenantCreateR
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.TenantCreateRequest) error); ok {
 		r1 = rf(_a0, _a1)
 	} else {
@@ -42,6 +45,10 @@ func (_m *TenantServiceServer) Delete(_a0 context.Context, _a1 *v1.TenantDeleteR
 	ret := _m.Called(_a0, _a1)
 
 	var r0 *v1.TenantResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantDeleteRequest) (*v1.TenantResponse, error)); ok {
+		return rf(_a0, _a1)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantDeleteRequest) *v1.TenantResponse); ok {
 		r0 = rf(_a0, _a1)
 	} else {
@@ -50,7 +57,6 @@ func (_m *TenantServiceServer) Delete(_a0 context.Context, _a1 *v1.TenantDeleteR
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.TenantDeleteRequest) error); ok {
 		r1 = rf(_a0, _a1)
 	} else {
@@ -65,6 +71,10 @@ func (_m *TenantServiceServer) Find(_a0 context.Context, _a1 *v1.TenantFindReque
 	ret := _m.Called(_a0, _a1)
 
 	var r0 *v1.TenantListResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantFindRequest) (*v1.TenantListResponse, error)); ok {
+		return rf(_a0, _a1)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantFindRequest) *v1.TenantListResponse); ok {
 		r0 = rf(_a0, _a1)
 	} else {
@@ -73,7 +83,6 @@ func (_m *TenantServiceServer) Find(_a0 context.Context, _a1 *v1.TenantFindReque
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.TenantFindRequest) error); ok {
 		r1 = rf(_a0, _a1)
 	} else {
@@ -88,6 +97,10 @@ func (_m *TenantServiceServer) Get(_a0 context.Context, _a1 *v1.TenantGetRequest
 	ret := _m.Called(_a0, _a1)
 
 	var r0 *v1.TenantResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantGetRequest) (*v1.TenantResponse, error)); ok {
+		return rf(_a0, _a1)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantGetRequest) *v1.TenantResponse); ok {
 		r0 = rf(_a0, _a1)
 	} else {
@@ -96,7 +109,6 @@ func (_m *TenantServiceServer) Get(_a0 context.Context, _a1 *v1.TenantGetRequest
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.TenantGetRequest) error); ok {
 		r1 = rf(_a0, _a1)
 	} else {
@@ -111,6 +123,10 @@ func (_m *TenantServiceServer) GetHistory(_a0 context.Context, _a1 *v1.TenantGet
 	ret := _m.Called(_a0, _a1)
 
 	var r0 *v1.TenantResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantGetHistoryRequest) (*v1.TenantResponse, error)); ok {
+		return rf(_a0, _a1)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantGetHistoryRequest) *v1.TenantResponse); ok {
 		r0 = rf(_a0, _a1)
 	} else {
@@ -119,7 +135,6 @@ func (_m *TenantServiceServer) GetHistory(_a0 context.Context, _a1 *v1.TenantGet
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.TenantGetHistoryRequest) error); ok {
 		r1 = rf(_a0, _a1)
 	} else {
@@ -134,6 +149,10 @@ func (_m *TenantServiceServer) Update(_a0 context.Context, _a1 *v1.TenantUpdateR
 	ret := _m.Called(_a0, _a1)
 
 	var r0 *v1.TenantResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantUpdateRequest) (*v1.TenantResponse, error)); ok {
+		return rf(_a0, _a1)
+	}
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantUpdateRequest) *v1.TenantResponse); ok {
 		r0 = rf(_a0, _a1)
 	} else {
@@ -142,7 +161,6 @@ func (_m *TenantServiceServer) Update(_a0 context.Context, _a1 *v1.TenantUpdateR
 		}
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(context.Context, *v1.TenantUpdateRequest) error); ok {
 		r1 = rf(_a0, _a1)
 	} else {
