@@ -3,14 +3,14 @@ package datastore
 import (
 	"context"
 	"errors"
-	"fmt"
 	"os"
 	"path"
 	"path/filepath"
-	"reflect"
 	"regexp"
 	"strings"
+	"time"
 
+	"github.com/jmoiron/sqlx"
 	healthv1 "github.com/metal-stack/masterdata-api/api/grpc/health/v1"
 	v1 "github.com/metal-stack/masterdata-api/api/v1"
 	"github.com/metal-stack/masterdata-api/pkg/health"
@@ -18,20 +18,51 @@ import (
 	"sigs.k8s.io/yaml"
 )
 
+type bootstrap[E Entity] struct {
+	log *zap.SugaredLogger
+	ds  Storage[E]
+}
+
 // Initdb reads all yaml files in given directory and apply their content as initial datasets.
-func (ds *Datastore) Initdb(healthServer *health.Server, dir string) error {
+func Initdb(log *zap.SugaredLogger, db *sqlx.DB, healthServer *health.Server, dir string) error {
+	start := time.Now()
 	files, err := filepath.Glob(path.Join(dir, "*.yaml"))
 	if err != nil {
 		return err
 	}
+
+	ts, err := New(log.Desugar(), db, &v1.Tenant{})
+	if err != nil {
+		return err
+	}
+	tbs := &bootstrap[*v1.Tenant]{
+		log: log,
+		ds:  ts,
+	}
+
+	ps, err := New(log.Desugar(), db, &v1.Project{})
+	if err != nil {
+		return err
+	}
+	pbs := &bootstrap[*v1.Project]{
+		log: log,
+		ds:  ps,
+	}
 	for _, f := range files {
-		ds.log.Info("read initdb", zap.Any("file", f))
-		err = ds.processConfig(f)
+		log.Infow("read initdb for tenants", "file", f)
+		err = tbs.processConfig(f)
 		if err != nil {
 			return err
 		}
 	}
-
+	for _, f := range files {
+		log.Infow("read initdb for projects", "file", f)
+		err = pbs.processConfig(f)
+		if err != nil {
+			return err
+		}
+	}
+	log.Infow("done reading initdb files", "took", time.Since(start))
 	healthServer.SetServingStatus("initdb", healthv1.HealthCheckResponse_SERVING)
 	return nil
 }
@@ -60,7 +91,7 @@ func splitYamlDocs(doc string) []string {
 }
 
 // processConfig processes all yaml docs contained in the given file
-func (ds *Datastore) processConfig(file string) error {
+func (ds *bootstrap[E]) processConfig(file string) error {
 	yml, err := os.ReadFile(file)
 	if err != nil {
 		return err
@@ -78,7 +109,7 @@ func (ds *Datastore) processConfig(file string) error {
 	return nil
 }
 
-func (ds *Datastore) createOrUpdate(ctx context.Context, ydoc []byte) error {
+func (bs *bootstrap[E]) createOrUpdate(ctx context.Context, ydoc []byte) error {
 
 	// all entities must contain a meta, parse that to get kind and apiversion
 	var mm MetaMeta
@@ -86,71 +117,64 @@ func (ds *Datastore) createOrUpdate(ctx context.Context, ydoc []byte) error {
 	if err != nil {
 		return err
 	}
-	ds.log.Info("initdb", zap.Any("meta", mm.Meta.GetKind()))
+	bs.log.Infow("initdb", "meta", mm.Meta.GetKind())
 
 	kind := mm.Meta.GetKind()
 	apiversion := mm.Meta.GetApiversion()
 
-	// get type for this kind from datastore entity types registry
-	typeElem, ok := ds.types[strings.ToLower(kind)]
-	if !ok {
-		return fmt.Errorf("initdb: unknown kind:%s", kind)
-	}
-	// messy extraction of apiversion from type
-	// type is like "*v1.Project"
-	elementType := reflect.TypeOf(typeElem)
-	expectedAPI := strings.ReplaceAll(strings.Split(elementType.String(), ".")[0], "*", "")
-	if expectedAPI != apiversion {
-		ds.log.Error("initdb apiversion does not match", zap.String("given", apiversion), zap.String("expected", expectedAPI))
+	var e E
+	if kind != e.Kind() {
+		bs.log.Infow("skip", "kind from yaml", kind, "required kind", e.Kind())
 		return nil
 	}
-	ds.log.Info("initdb", zap.Stringer("type", elementType), zap.String("apiversion", apiversion))
 
-	// no we have the type, create new from type and marshall in that new struct
-	newEntity, ok := reflect.New(elementType.Elem()).Interface().(VersionedJSONEntity)
-	if !ok {
-		panic(fmt.Sprintf("entity type %s must implement VersionedJSONEntity-Interface", elementType.String()))
+	// messy extraction of apiversion from type
+	if e.APIVersion() != apiversion {
+		bs.log.Errorw("initdb apiversion does not match", "given", apiversion, "expected", e.APIVersion())
+		return nil
 	}
 
-	err = yaml.Unmarshal(ydoc, newEntity)
+	ee := new(E)
+	err = yaml.Unmarshal(ydoc, ee)
 	if err != nil {
 		return err
 	}
+	e = *ee
 
-	newKind := newEntity.GetMeta().GetKind()
-	newID := newEntity.GetMeta().GetId()
+	newKind := e.GetMeta().GetKind()
+	newID := e.GetMeta().GetId()
 
 	// now check that this type is already present for this id,
 	// therefore create nil interface to get into
-	existingEntity := reflect.New(elementType.Elem()).Interface().(VersionedJSONEntity)
-	err = ds.Get(ctx, mm.Meta.GetId(), existingEntity)
+	exists := true
+	existingEntity, err := bs.ds.Get(ctx, mm.Meta.GetId())
 	if err != nil {
 		if errors.As(err, &NotFoundError{}) {
-			existingEntity = nil
+			exists = false
 		} else {
-			ds.log.Error("initdb", zap.Error(err))
+			bs.log.Errorw("initdb", "error", err)
 			return err
 		}
 	}
 	// now check if it exists by checking for id presence
 	// then update, otherwise create
-	if existingEntity != nil {
+	if exists {
 		oldVersion := existingEntity.GetMeta().GetVersion()
-		newVersion := newEntity.GetMeta().GetVersion()
-		ds.log.Info("initdb found existing, update", zap.String("kind", newKind), zap.String("id", newID), zap.Any("old version", oldVersion), zap.Any("new version", newVersion))
+		newVersion := e.GetMeta().GetVersion()
+		bs.log.Infow("initdb found existing, update", "kind", newKind, "id", newID, "old version", oldVersion, "new version", newVersion)
 		if oldVersion >= newVersion {
-			ds.log.Info("initdb existing version is equal or higher, skip update", zap.String("kind", newKind), zap.String("id", newID))
+			bs.log.Infow("initdb existing version is equal or higher, skip update", "kind", newKind, "id", newID)
 			return nil
 		}
 
-		newEntity.GetMeta().SetVersion(existingEntity.GetMeta().GetVersion())
-		err = ds.Update(ctx, newEntity)
+		e.GetMeta().SetVersion(existingEntity.GetMeta().GetVersion())
+		err = bs.ds.Update(ctx, e)
 		if err != nil {
 			return err
 		}
 	} else {
-		ds.log.Info("initdb create", zap.String("kind", newKind), zap.String("id", newID))
-		err = ds.Create(ctx, newEntity)
+		bs.log.Infow("initdb create", "kind", newKind, "id", newID)
+		err = bs.ds.Create(ctx, e)
 		if err != nil {
 			return err
 		}
