@@ -1,4 +1,4 @@
-// Code generated by mockery v2.14.0. DO NOT EDIT.
+// Code generated by mockery v2.24.0. DO NOT EDIT.
 
 package mocks
 
@@ -7,13 +7,13 @@ import (
 	mock "github.com/stretchr/testify/mock"
 )
 
-// VersionedJSONEntity is an autogenerated mock type for the VersionedJSONEntity type
-type VersionedJSONEntity struct {
+// Entity is an autogenerated mock type for the Entity type
+type Entity struct {
 	mock.Mock
 }
 
 // APIVersion provides a mock function with given fields:
-func (_m *VersionedJSONEntity) APIVersion() string {
+func (_m *Entity) APIVersion() string {
 	ret := _m.Called()
 
 	var r0 string
@@ -27,7 +27,7 @@ func (_m *VersionedJSONEntity) APIVersion() string {
 }
 
 // GetMeta provides a mock function with given fields:
-func (_m *VersionedJSONEntity) GetMeta() *v1.Meta {
+func (_m *Entity) GetMeta() *v1.Meta {
 	ret := _m.Called()
 
 	var r0 *v1.Meta
@@ -43,7 +43,7 @@ func (_m *VersionedJSONEntity) GetMeta() *v1.Meta {
 }
 
 // JSONField provides a mock function with given fields:
-func (_m *VersionedJSONEntity) JSONField() string {
+func (_m *Entity) JSONField() string {
 	ret := _m.Called()
 
 	var r0 string
@@ -57,7 +57,7 @@ func (_m *VersionedJSONEntity) JSONField() string {
 }
 
 // Kind provides a mock function with given fields:
-func (_m *VersionedJSONEntity) Kind() string {
+func (_m *Entity) Kind() string {
 	ret := _m.Called()
 
 	var r0 string
@@ -71,7 +71,7 @@ func (_m *VersionedJSONEntity) Kind() string {
 }
 
 // Schema provides a mock function with given fields:
-func (_m *VersionedJSONEntity) Schema() string {
+func (_m *Entity) Schema() string {
 	ret := _m.Called()
 
 	var r0 string
@@ -85,7 +85,7 @@ func (_m *VersionedJSONEntity) Schema() string {
 }
 
 // TableName provides a mock function with given fields:
-func (_m *VersionedJSONEntity) TableName() string {
+func (_m *Entity) TableName() string {
 	ret := _m.Called()
 
 	var r0 string
@@ -98,14 +98,14 @@ func (_m *VersionedJSONEntity) TableName() string {
 	return r0
 }
 
-type mockConstructorTestingTNewVersionedJSONEntity interface {
+type mockConstructorTestingTNewEntity interface {
 	mock.TestingT
 	Cleanup(func())
 }
 
-// NewVersionedJSONEntity creates a new instance of VersionedJSONEntity. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewVersionedJSONEntity(t mockConstructorTestingTNewVersionedJSONEntity) *VersionedJSONEntity {
-	mock := &VersionedJSONEntity{}
+// NewEntity creates a new instance of Entity. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+func NewEntity(t mockConstructorTestingTNewEntity) *Entity {
+	mock := &Entity{}
 	mock.Mock.Test(t)
 
 	t.Cleanup(func() { mock.AssertExpectations(t) })
